[roll] Roll fuchsia [settings] Warn on missing registrant dependency
Rather than silently skipping the registration of a `Registrant` that is
missing a dpendency we now log that the registration is being skipped.
The name of the registrant is now tracked so that it can be reported
when there is an error, to support that some cleanup was done as well.

Tests: fx setui_service_tests
Original-Fixed: 96251
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/784442
Original-Revision: f425b13fa5c31bf9e214142f03cbfbd5b42b93c8
GitOrigin-RevId: 736d696e619a600450135a2656b735bf2c5bccbe
Change-Id: I4bb5b4cee519b4317b501d8a7e5dd51982e3da48