[roll] Roll fuchsia [netemul] Keep strong ref to realm from Interface
TestInterface had to eagerly connect to a number of services in its
realm so fulfill its helper functions.

By having access to its containing realm we can connect to the services
when needed.

Ideally, TestInterface would keep a borrow of the realm so that we can
encode the lifetimes, but that makes helper functions that returns
TestRealm + TestInterface impossible because that would be a
self-referential return.

Prior to this change, destroying a realm before the TestInterface would
cause most of the helpers to fail regardless, because the realm
containing the netstack backing the services would be tearing down.

With this change, TestInterface clones the ManagedRealm proxy, keeping
effectively a clone of the TestRealm for itself, which allows it to
connect to services on demand. An unexplored benefit is that helper
functions might drop TestRealms if they only care about a single
interface installed in that realm.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/784303
Original-Revision: b00dc33759c484dda1d066ddcfbae10cbcf883e1
GitOrigin-RevId: a14d70e0b113972a7c25c2cfe5aa462e5f944903
Change-Id: Ice64d462b705ca3c348d408f9432b6d812fd1efc