[roll] Roll fuchsia [media][codec] Implement roundtrip test for CVSD.
Implement test suite for doing roundtrip testing for codecs. It performs
the roundtrip test by calculating the target RMSE from golden files and
running the input through the encoder and the decoder. It compares the
resulting RMSE from the codec-under-test against the target RMSE to
ensure that the RMSE is within the limit. It also performs one-way
test using just the golden encoded file and the decoder to ensure that
the codec works as intended.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759303
Original-Revision: f72e24585c03669f2efff90994764e86e21de4a0
GitOrigin-RevId: aaa41b96d37658032f8fa7827c48678c629cef3d
Change-Id: I63b7d16e26b43fdd5701205caecd8e880814b2f1