[roll] Roll infra/recipes [btu] Add option to pass Jiri attributes
This is a checkout optimization intended to reduce the number of CIPD
packages downloaded which are typically unnecessary to build the common
configurations.

Note that Jiri is on its way out and this is only a temporary solution
until we can use `bazel fetch` to fetch only the packages needed by the
build graph prior to running `bazel build`.

Original-Bug: 117646
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/785834
Original-Revision: c699d11d461712a5bd8d235abc2871aea2b8a2f9
GitOrigin-RevId: f47c582277ae27f4aa96d452a1bba608f8b3e6e0
Change-Id: Ie9ef6f07eefc167242a5be91ae1f18ee699d7fd7