[roll] Roll fuchsia [moniker] Make &str -> ChildMoniker fallible
We implemented `From<&str>` for ChildMoniker and unwrapped the
result implicitly. This was mostly used in tests, but is a
dangerous implementation because it is also available for
production use.

The Rust Language also specifies that the From trait should
only be used for infallible conversions.

This CL implemented `TryFrom<&str>` for ChildMoniker and
changes all tests to unwrap explicitly.

Original-Bug: 118649
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786022
Original-Revision: c7dad8aec73bfb3648c94f47261a45dd9aafec90
GitOrigin-RevId: 107577d623b8cd864d9c144a0af65c0df682a585
Change-Id: I84083de5418432fac0e79be1aa55cf74e519e97e