[roll] Roll fuchsia [moniker] Make &str -> InstancedChildMoniker fallible
We implemented `From<&str>` for InstancedChildMoniker and
unwrapped the result implicitly. This was mostly used in
tests, but is a dangerous implementation because it is also
available for production use.

The Rust Language also specifies that the From trait should
only be used for infallible conversions.

This CL implemented `TryFrom<&str>` for InstancedChildMoniker
and changes all tests to unwrap explicitly.

Original-Bug: 118649
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786082
Original-Revision: ff223396bf6f2bc0d9ccc677fd4bb1c92d2fbf27
GitOrigin-RevId: f4895d7b3ccf16db41549dcb4a2bfba92710bfe0
Change-Id: I76f38df1e392884b6277873a92324dc3175e4ba3