[roll] Roll fuchsia [settings] Allow missing camera on restore
Rather than failing to bring up the service when the camera is missing,
allow the service to pass restore. In this state the camera status
should show that it's in an error state. Allow clients to attempt to
update the state of the camera, since it could reconnect and bring the
camera out of the error state. If the camera is still stuck, the state
will still show as an error in future requests.

Original-Bug: 117854
Test: fx test -o setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785510
Original-Revision: 94a5b21294bad0cebcffa1c28412ac38ccec0208
GitOrigin-RevId: 16286ef5fa2d35a7373809f87b5745cd73c3a63b
Change-Id: Iee9af6c2a6cb83d89aed5764770a13695d590ae1