[roll] Roll fuchsia [elfldltl] Clean up VMO name handling in VmarLoader
The SetVmoName helper doesn't need to be a method and so doesn't
need to be declared in the public header.  Regardless, the
constants used as template parameters don't need to be public
since they're not themsleves used from code in the header.

Original-Bug: 91206
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785746
Original-Revision: a0d8108ae948bd634ee42dd6bf3e51cd0ca61390
GitOrigin-RevId: 3e3994591b3738f91a98ae7a4ef40c9ee2509561
Change-Id: Iedd84329ae371f9f5fb903dd9d911751e6d1c545