[roll] Roll fuchsia [ffx][driver] Add `--fail-on-missing` flag to `ffx driver list-devices`
This makes it easier to use `ffx driver list-devices` on shell scripts
to detect the presence of a certain device and fail otherwise.

Preferably we would always silently return a non-zero exit code when no
matching drivers are found, but FFX will print an error message and a
backtrace if a subcommand returns an error, which is not desirable as
the default for this case.

Original-Bug: 113465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781346
Original-Revision: 22a058b1a228896deca643fec78132e5eb2dc812
GitOrigin-RevId: 1a422cd32d4dae8a57b35bb8cd54b032071c4804
Change-Id: Ia279ad836dc5a019fa90d3eae307195ce300835d