[roll] Roll fuchsia [docs] Include sections on ffx repository and package
We are in the process of migrating of `pm` over to `ffx package` and
`ffx repository`. This updates the documentation to include sections
that describe how to use the new tools and the legacy `pm` that will
stick around until we've completed the migration.

Original-Bug: 116114
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781582
Original-Revision: ad4882b34bdfa55d35b9c20b05b00953fc9e60e4
GitOrigin-RevId: 2e553af1d6e3d08c679856136f255ef41171137c
Change-Id: I169ba5646aa836b1b23a6656f28b8893c27d659f