[roll] Roll fuchsia [component_manager] Fix leak in events synthesis
Events synthesis spawns a task with `detach` that performs the
depth-first exploration for event generation. Bind the task to the
target component's TaskScope so it gets released when the component is
destroyed.

This leak has existed for a long time, but it was exacerbated by
fxrev.dev/761894 since that it makes it inevitable that event routing
will complete.

Original-Bug: 116722
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780647
Original-Revision: 27c2fc281f0f0dadd9ae2bd113aa49f0f4b5b566
GitOrigin-RevId: d71087f7ac015f9f22348b7e9f96fac6757db3e2
Change-Id: I9d950566426492e267230b3e3d0297cbe34350a0