[roll] Roll fuchsia [fdf] Update DFv1 composite device workflow
Make changes to the DFv1 composite device workflow
so that a node representing the composite is
only added to the node topology when there's a
driver bound to it.

CompositeDevice now stores a |driver_| object, which
is set by three methods:
1) If CompositeDevice is created by
   CreateFromDriverIndex().
2) CompositeDevice::TryAssemble() - if this function
   is called and |driver_| is nullopt, then
   CompositeDevice will try to find a matching driver
   to set |driver_|.
3) BindDriverManager::BindAllDevices() - when this
   function is called, it'll try to bind a matching
   driver to all devices and CompositeDevice objects.
   This is so if a matching driver is loaded after
   CompositeDevice::TryAssemble() was called, the
   driver can now bind to actual device.

This CL also makes changes to device_enumeration and
src/devices/driver_manager tests.

For device_enumeration Sherlock test, aml-hevc-enc is
removed since the driver that binds to it is
excluded in bringup.

In driver_manager tests, MultipleDeviceTestCase is
modified so that creating the Coordinator config is
overridable by subclasses. Some common code are also
refactored in CompositeDeviceTests.

Original-Bug:118642
Tests: Update driver_manager tests. Added test to
ensure the device is only created when it has a
matching driver

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785483
Original-Revision: 91cfed6078017bffac23439c71232a7633725599
GitOrigin-RevId: 8e001f20772cd35ce6d95fc1f2f7a6fceac1c7eb
Change-Id: I58855eaffa5be08a66a91a42e7ee15e17b7f3bc2