[roll] Roll fuchsia [moniker] Make Vec<&str> -> AbsoluteMoniker fallible
We implemented `From<Vec<&str>>` for AbsoluteMoniker and
unwrapped the result implicitly. This was mostly used in
tests, but is a dangerous implementation because it is also
available for production use.

The Rust Language also specifies that the From trait should
only be used for infallible conversions.

This CL implemented `TryFrom<Vec<&str>>` for AbsoluteMoniker
and changes all tests to unwrap explicitly.

Original-Bug: 118649
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786362
Original-Revision: a8e7f1423f5c85ee5cd8d53811e0c92429f46c16
GitOrigin-RevId: e742340dfe95f388e5f32834057b31478d73a0f7
Change-Id: Ic9d49b556b1cf4a46549349ffb8d3a704c6b976d