[roll] Roll fuchsia [moniker] Make Vec<&str> -> InstancedAbsoluteMoniker fallible
We implemented `From<Vec<&str>>` for InstancedAbsoluteMoniker
and unwrapped the result implicitly. This was mostly used in
tests, but is a dangerous implementation because it is also
available for production use.

The Rust Language also specifies that the From trait should
only be used for infallible conversions.

This CL implemented `TryFrom<Vec<&str>>` for
InstancedAbsoluteMoniker and changes all tests to unwrap
explicitly.

Original-Fixed: 118649
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786502
Original-Revision: 194b9e18e4d9d60a7a803d821828fb03af027884
GitOrigin-RevId: 0f02da4d6820a34dd16e59198c26300f468813c0
Change-Id: I8ce5e5d984ae7abb71e2fa1e190be96627ac5111