[roll] Roll fuchsia [cargo-gnaw] Add support for building unit test targets.
This CL adds support for generating executable targets for each library
and binary with tests. We only generate tests for targets that
explicitly enable it in the root Cargo.toml.

This CL does not attempt to deal with hooking these tests up to the rest
of the build.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765965
Original-Revision: f055e4db396879bcbfb244a453fc948af095d90a
GitOrigin-RevId: 8e067227cbda10786d7c9a66794f69c9b920c861
Change-Id: Ia73ddbfd4f8059f5afb7db18a4b7fb3fee94d682