[roll] Roll fuchsia [elfldltl] Disable hwasan on static globals in memory test
Hwasan reports a tag mismatch for globals accessed via the memory API
since the tag for the image start will almost never be the same as the
tag for an arbitrary global within that image. The test is WAI, but just
isn't compatible with hwasan.

Original-Fixed: 118713
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786555
Original-Revision: 458413eb712f7591841046865aa82287cd3aecf0
GitOrigin-RevId: 5cd2c756351a661c39f152111f622dd7dad734cc
Change-Id: I655da8d41befb418848ac8910ff0682569653687