[roll] Roll fuchsia [kernel][vm] Remove incorrect can_borrow_locked assert
The value of is_any_borrowing_enabled() can change between when it is
used to evaluate the test condition and the assert. Remove the assert
and replace it with a check for debug_is_user_pager_backed() which is
the only non-redundant portion of the assert anyway.

Original-Fixed: 118821
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786428
Original-Revision: 4a3f8962a017201393069aea1e8fdbb52a82fefd
GitOrigin-RevId: 02456ba7795c5ac4deacc1bb86c5d5e8550c019a
Change-Id: I2b3600abcf6db7f7d81d30012e9b4466ea7f4966