[roll] Roll fuchsia [netstack3] Factor out multi-nic network setup
Factor out a helper that sets up hosts in a star topology
where the multi-nic host is connected to each peer through a separate
network, and none of the peers are connected to each other. This is
useful for testing that sockets can be bound to devices explicitly or
implicitly via zoned addressing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785563
Original-Revision: bf0584c74654d84ea8a69bd37c87ab27f1713202
GitOrigin-RevId: e9e3e0cc1da2050c231daf84787531e70c1232cc
Change-Id: Ie9145345f37e500baf36113974bf6bfd43e9cde6