[roll] Roll fuchsia [fshost] Don't file crash reports on FDR
When Fxfs is using the legacy crypto format, we implement FDR by
creating a marker file in Fxfs which is detected on boot and which
results in fshost reformatting Fxfs.  This also creates a crash report
at the moment, which is not what we want.

Fix this to not file a crash report.  Also fixes the integration tests
to actually exercise this (the relevant tests were matching for
the "fxfs" feature, but the integration tests for the legacy format set
the "fxfs-legacy-crypto" feature instead), since they would have caught
this bug.

Original-Fixed: 118909
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787090
Original-Revision: f7aab0a98a16c715a1e831eecf84ff7f86573091
GitOrigin-RevId: 3c1e315070cf327dde2f91ce121c3b508ee81023
Change-Id: I3153985e3f745bb5c73ce6b97dfe1b322ce62916