[roll] Roll fuchsia [fdf] Check rebind during IsUnbound
Check that `pending_rebind_` is false during the compat device's
`IsUnbound()`. The dev connector closes a connection that attempts to
connect to a device's controller if `IsUnbound()` is true.
`IsUnbound()` used to only check if the device is pending removal.  If a
DFv1 device in a DFv2 compat shim is rebound then the compat shim marks
the device for pending removal. Now if a DFv1 device in a DFv2 compat
shim is rebound, the dev connector will successfully connect to the
device's controller instead of closing the connection.

Set `pending_removal_` to false once a compat device has been rebound.

Bug 117669

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786742
Original-Revision: 932fb6efedc0ab9bb559cdb8d29116833bbb0970
GitOrigin-RevId: fe86aad382f412800b2a33d62861a3c3f634d578
Change-Id: I6ba44209f508474f4cde9be6c357e4f2fe3c502b