[roll] Roll fuchsia [device-watcher] Fix request race with teardown
Previously, `IterateDirectoryTest` tests would create requests to the
fake filesystem via the `device-watcher` library's `IterateDirectory()`
function. The requests are asynchronous/pipelined and not confirmed to
be acknowledged by the server by the time it is passed to the test's
callback. Some tests would not verify that the connection is
acknowledged before finishing, which would cause the fake filesystem and
its async loop to teardown as well. If the destruction occurs before the
outgoing request is acknowledged, this would cause the request to be
dispatched on a dead async loop.

This change uses `ManagedVfs` instead, which can handle this case.

Original-Fixed: 118686
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782024
Original-Revision: c963a9aaab146443a791ee8d4ba420f8a22389df
GitOrigin-RevId: 76a448a036d68fc2f773b278fd21f1c2adbafe73
Change-Id: I72043294972ca00f725dd152b1b87a90920b80c9