[roll] Roll fuchsia [fidlgen_rust] Derive Copy/Clone based on resourceness
This CL makes it so that resource types in Rust never derive Copy and
Clone, even when they have no handles. This was the intention all along:
marking a type as `resource` is an up front decision that means adding a
handle later will not be source breaking.

Note that the presence of handles is still used for deriving zerocopy
traits, since these are meant to be an encoding implementation detail.
(Technically this could cause source breakage if someone used them, but
that already happens with vectors and Copy or floats and Eq/Ord/Hash).

Added an empty resource struct, table, and union in
handles_in_types.test.fidl to demonstrate.

Also refactored the derives algorithm to make it simpler.

Original-Fixed: 118784
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786438
Original-Revision: 8b03e84a6ec85e08d4bc32c6155c9efc0c58d170
GitOrigin-RevId: 462d8df42ee2080698221410fda51990cc809c6a
Change-Id: I55515fb99aef8a0398123d9388f9a4a52f6f3751