[roll] Roll fuchsia [virtualization] Route fuchsia.input.wayland.Keymap virtio_wl.
Now that the wayland server is inside the virtio_wl device, we need to
update the routing for this service. The addition of this service
conflicted with some of my other changes so this is just fixing up
routing for this service, both for static-child vmm's and also for vmms
launched using the newer vmm_launcher path.

Original-Bug: 115651
Test: Observed the routing error for wayland.Keymap disappears when
      booting the linux terminal on workstation.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787203
Original-Revision: 23ec5c30263d2ca706af58d211f8b8317d23f0eb
GitOrigin-RevId: d79447f696ed162ed96c2d27e7df8de1b7094dc2
Change-Id: I4aa3050f19722b441b315b6794b42320dde9428c