[roll] Roll third_party/antlion Fix DHCP options error in Dhcpv4InteropTest
The Dhcpv4InteropTest was updated to use generate_tests but does not
have an option for test_name which was added. This change fixes that bug
and refactors the test to improve the readability of the test
parameters.

Original-Fixed: b/232574848
Test: Dhcpv4InteropTest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/787282
Original-Revision: 73965b78a0f8f71ae56aa206b346e44e7e315f3d
GitOrigin-RevId: 5d1d1d5051b35ddc5a2e791939cb28613603dbc8
Change-Id: Ia34480b0672928ae96e1cdd576a8ec328f23c914