[roll] Roll fuchsia [kernel][ktrace] Migrate to interned categories.
This change is an intermediate migration to FXT interned categories in
the kernel and ktrace. To maintain compatibility with the existing
ktrace syscall interface, the defined ktrace filter groups are
statically mapped to interned categories with the same category names
that ktrace_provider uses.

- Update ktrace session logic to test categories instead of tags.
- Update ktrace fxt wrappers to accept categories instead of tags.
- Update trace points to supply categories instead of tags.
- Update category checks and remove redundancy.

Test: Observe traces with various settings
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780749
Original-Revision: 57f74dea2b9b997b58ffaddacdd6aa389771e026
GitOrigin-RevId: 391cda742707da3aaa36207a148a71cf7d8b4c51
Change-Id: I6649862d785cf05cc35c5d030bbb175a16540ed6