[roll] Roll fuchsia [fasync] Move on_closed() to fasync::Channel
This CL implements on_closed() on fasync::Channel (zircon and emulated)
so that FIDL's Proxy::on_closed() can just delegate to it. This is in
preparation for adding on_closed() for servers in FIDL as well. And it
seems like it belongs more on Channel anyway.

Original-Bug: 101155
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787057
Original-Revision: 1546aae6b90b7d51598aa88fe541696d981ce68d
GitOrigin-RevId: 80120ad25fa3865e1932cf718d97837062dbd9b1
Change-Id: Iedd673224ba82b51b166b3db64814530b50a6f35