[roll] Roll fuchsia [fdf] Reland Update DFv1 composite device workflow
This is a reland of
commit 91cfed6078017bffac23439c71232a7633725599.

To prevent the CompositeDeviceTests from hanging,
we now shutdown the FakeDriverIndex's thread.

In addition, error handling in BindAllDevices() is
adjusted so that it'll continue to match and bind
the other composites even when it encounters an
error.

Multiply: fuchsia-pkg://fuchsia.com/driver-manager-tests#meta/driver-manager-test.cm : 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786848
Original-Revision: e041f6b3f67cd688eb6cf6226c5c03afb47bb8d0
GitOrigin-RevId: 4c740eccb5b46c3f87fd8587aa5664883c87656a
Change-Id: I49a80b05c559c9bb7985c27827ce8be9d4815812