[roll] Roll fuchsia [helpdoc] Fixing README path initialization
This fixes a problem initializing the README path for the categories
of tools. The issue was caused by incorrectly organizing the local
variables, causing one of them to be initialized based on another
variable which was not initialized yet.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787217
Original-Revision: a51c8830bf9070642fd0b086a55d7c1a67e2b982
GitOrigin-RevId: c203644a8b4f86d134dbba671baf3c301c935a55
Change-Id: I3b10f4528108c566dfa4bda13f426b3cc2eadd22