[roll] Roll fuchsia [wlan][policy] fix scoring log spam in match
There was a typo in a match statement that lead to a lot of log spam if
the value matched the intended bucket. This changes the match style of
the match statements to be less likely to have similar typos, and there is
no unexpected case that would cause logging anymore.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787508
Original-Revision: d3a84a2ee73078740fc6a51c74492296c694f0a8
GitOrigin-RevId: e9d479b54468eb3d0c6e1d05482d7962b7428612
Change-Id: I48b9b808dde3552b83681def8631037ba850e2b5