[roll] Roll fuchsia [nvme] Fix driver shutdown deadlock and re-enable tests.
In Namespace::IoLoop(), the driver shutdown boolean is checked in
between waiting for a sync completion and resetting it. Hence, it's
possible for both the driver shutdown boolean to be set and the
sync completion to be signalled (in Namespace::DdkRelease()) in the
narrow time window between after checking the driver shutdown
boolean but before resetting the sync completion, which results in
a deadlock.

The cancellation of pending IOs at driver shutdown time are handled
in Namespace::DdkRelease().

Original-Bug: 118907
Multiply: fuchsia-pkg://fuchsia.com/nvme-tests#meta/nvme-unit-tests.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787307
Original-Revision: adedd13bee7f60284941b4ae15ddc4b0925e03be
GitOrigin-RevId: 8c5fded371ee35f3828299ea5aaef3f15640029a
Change-Id: I8f86ec89263e32dcdeaf6ecb83179f9e0f1737e7