[roll] Roll fuchsia [emu] Enable custom virtual devices for emu
This CL adds logic to the handling of the "--device" flag to allow the
user to specify a custom virtual device file path. If the input to
"--device" is not a path to a file, `ffx emu` will continue to treat it
as the name of a virtual device embedded in the target product bundle.

Original-Bug: 108747
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778324
Original-Revision: 13071925408bcf937770b39ea678b6c2a9c9962b
GitOrigin-RevId: e01982d9591d8af40413cebad5e16a19ff2ccab1
Change-Id: I7ea373ccd4d223a2d7cbc915cbc187de0611cb00