[roll] Roll fuchsia [runtime] Start with 0 threads
Start the runtime with 0 threads, and provide an environment function to
start the driver runtime with it's first thread.

- DriverHost 1 and 2 are updated to start the driver runtime.
- driver-runtime-main is updated to start the driver runtime.
- A new gtest version of driver-runtime-main is created.
- Most driver runtime tests are updated to start the driver runtime.

This change will make driver unittesting much simpler, as tests by
default will not need to deal with threading concerns.

A new test function (fdf_testing_run_until_idle) is added which lets the
test run the main loop until it is idle.

The DFv2 driver unittests will be updated to make use of this in a
followup CL.

Original-Bug: 103368

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787123
Original-Revision: 4997d37dc5c1439964ed388af5fb0f5afb4c8475
GitOrigin-RevId: 507f7eb0e3528d5046f8206a5ac5b5e81ccc52bd
Change-Id: Id6a90f6742c6ddbaa30de6a4a870d4a6df1d460f