[roll] Roll fuchsia [starnix] Extend MAP_GROWSDOWN mapping in response to user faults
This implements a portion of the MAP_GROWSDOWN feature that extends a
mapping to cover a faulting access from userspace. To support this the
memory manager has a feature to extend mappings for a single address
which is invoked when receiving a fault for a not-present page.

This should eventually be the only mechanism for growing such regions,
however since Starnix<->restricted copies currently happen through VMO
operations and not mapped memory they will need a separate codepath that
performs growth over a range of addresses to work properly. This change
also does not implement guard regions or the PROT_GROWSDOWN flag.

Original-Bug: 97514

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786430
Original-Revision: b90b92b3199d720716f05ab9298ca76fd0469f64
GitOrigin-RevId: ae34a6c2af5373ca22b4b1659c33762aa7c1244e
Change-Id: Ibdf44bf06f3e775262f8c97cd09540e1bb93723f