[roll] Roll gwp-asan [GWP-ASan] Fix atfork handlers being installed multiple times in tests
We incorrectly install the atfork handlers multiple times in the test
harness, tracked down to the default parameter used by
CheckLateInitIsOK. This manifested in a hang if running the tests with
--gtest_repeat={>=2} as the atfork handler ran multiple times, causing
double-lock and double-unlock, which on my machine hung.

Add a check-fail for this case as well to prevent this from happening
again (it was difficult to track down and is an easy mistake to make).

Differential Revision: https://reviews.llvm.org/D139731

GitOrigin-RevId: 0dd68f7b6ed5236187f4dba224e98dc6ce07b146
Original-Revision: ebca419199eb7fe574da5bc4975ddd0e4ccc2640
Change-Id: I3fbdc09c16afadaf62b0fec03c78057a017f4bd2