[roll] Roll fuchsia [emu] Fix --reuse semantics
The attached bug called out an unanticipated change in the semantics of
--reuse; when the instance did not already exist, the new semantics
would fail the launcher instead of creating the new instance.

This CL adds a new EngineOption type as output for get_engine_by_name,
such that callers can distinguish between failing to retrieve an
existing instance and that instance simply not existing. Now, if the
instance didn't exist before the start call with --reuse, we notify the
user, revert the --reuse flag so that the instance can be properly
initialized, and then continue the start sequence.

Original-Bug: 118006
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786441
Original-Revision: 3bf1bec8930695fb771385f1f064219d01de5db4
GitOrigin-RevId: 0371b0e3c145f1ece6ab71e112b150974d87d793
Change-Id: I3a678cf02ce082e21bfa19bc630df80d41fd45dc