[roll] Roll fuchsia [fbl] assert on BitField overflow
This commits promotes a ZX_DEBUG_ASSERT to a ZX_ASSERT in
BitFieldElement's operator=, disallowing setting values that are greater
than the maximum.

BitFieldElement operator= allows assigning a value to a BitField, and
there was a DEBUG_ASSERT that said value is less than the maximum possible
value for that element. However, in builds without debug asserts, the
code would still set the value, affecting bits of the bitfield unrelated
to the current element.

Test: fx test "fuchsia-pkg://fuchsia.com/fbl-test#meta/fbl-test.cm"
Original-Bug: 115898
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764255
Original-Revision: 623c079fbfead8b3427fcc4ddddffbac9720e55d
GitOrigin-RevId: 8ab1d559cabd17d5fa3a8c5186837b831c27f044
Change-Id: Ie3a1ce3c36fd34cb16756770c99c951de5cf9814