[roll] Roll fuchsia [chromium] Make canonical target empty
A canonical target is required per
https://fuchsia.dev/fuchsia-src/development/source_code/layout?hl=en#canonical_targets,
but nothing should depend on it for specific package(s). An empty
target addresses the requirement while not affecting product behavior
or needing to decide what it should include.

This removes one TODO for closed fxbug.dev/81925. Also update the
tracking issue for the other such TODO.

Move the canonical target to the BUILD.gn file for the directory where
it is needed.

Original-Bug: 119075, 81925, 115624
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786551
Original-Revision: 84c00eb78ca0bc2ae6de4707a1dc29847fbd7227
GitOrigin-RevId: d1ba2d55a13a68182ff3cd7b5fe3b74de7859b55
Change-Id: I86e4cae77489c8d41cb376f163eff139dd45c815