[roll] Roll fuchsia [gnsdk] Fix output manifest name for coverage test components
Recent changes to component.gni and prepare_package_outputs.py
resulted in the output name of the cm file for tests with coverage
enabled not matching the test name (it had a prefix and suffix added).

This adjusts it so that the output name of the merged manifest is
the same as the base manifest, but places it in a subdir to avoid
output path collisions.

Original-Bug: chromium:1406284
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788034
Original-Revision: fcb796064b412d974855bfb53565b452b81e6b8a
GitOrigin-RevId: 0997401677893a94bd9f8d7ef3e39d68eaabe3c1
Change-Id: Ic7daf8940a7c79c084412172fbda2cd99ea5fa6f