[roll] Roll fuchsia [ffx][product] add Fuchsia Tools client id and secret
This isn't a web app, so the "secret" is not secret. As is done for
gsutil and other client side executables, the client id and secret
are baked into the program in the source. This means that Refresh
tokens must be kept secret (they should be anyway, but for client
executables the security comes down to the refresh token).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787309
Original-Revision: 04f8a674aafb3eee85622faed1e85f23913f1fae
GitOrigin-RevId: cde5566a5d780cecc6ecd4fd627ecde6f4df9b64
Change-Id: If9c60ee09a5cd4107858a700fd23036e088ca482