[roll] Roll fuchsia [fidl] Add Fixable* helpers to Reporter
The |Reporter| now has helpers to report |Fixable[Error|Warn]|s, which
should be used to report errors that that should be reported as such
(ie, always build failing) in diagnostics, but can optionally be ignored
and disabled when fixing. This implies that compilation should be able
to fully recover from such errors - exiting out early on them is not
allowed, as there are scenarios where we want compilation to always
complete in spite of them.

Test: fx test fidlc-test
Original-Bug: 114357
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782896
Original-Revision: 38e63520a0b061ccf605756253f690ca90232e8b
GitOrigin-RevId: b9c369266ea31b28cdfa75748334f342684b3973
Change-Id: Ibe7f3bedb2b70ba5f823a7d303a54c2f5316f52b