[roll] Roll fuchsia [hierarchy] Remove usage of regexes in select_from_hierarchy
`select_from_hierarchy` was previously executing a `filter_hierarchy`.
In the case of Sampler, this wasn't necessary since Sampler receives
hierarchies that have been filtered by the Archivist.

In the case of triage, there's no need to filter the whole hierarchy and
then collect the properties into a vector, we can instead get the
properties as we traverse the hierarchy in a DFS way.

This still doesn't remove usage of RegexSet in the Archivist but it does
for Sampler.

Additionally the new implementation removes cloning of data and returns
the selected properties in a hierarchy as references to the original
hierarchy. There's some cloning still in triage that might be possible
to optimize into dealing with references, but that's out of scope of
this CL as Sampler is the main target.

This refactor brings back ~300K in Sampler:
https://screenshot.googleplex.com/WXqGbYVzNzV5FwN.png.

Original-Bug: 117157
Run-All-Tests: True

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/777862
Original-Revision: f0354b652c0439175a3a0b45b21649f4bbba0cb4
GitOrigin-RevId: ab0539a2c0379d1d0336b8862775513463fbb375
Change-Id: Ib134ef91c51da6d835fb55a780b06ad49aa923f5