[roll] Roll fuchsia [driver_manager] Another `has_outgoing_directory` -> `uses_fragment_driver`
Since we are specifically looking for a fragment that uses the fragment
driver, this is the appropriate check rather than whether the bound
device has an outgoing directory. These conditions used to be equivalent
but will not be in the future when we enable dual Banjo/FIDL nodes.

Original-Bug: 116754
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787902
Original-Revision: ac1b4e66136ed034eb3152de28e54a54dfd24e70
GitOrigin-RevId: 6632c4635bd9767f492752cc190ffe0c9553e193
Change-Id: I99f5b9acdc347e4cf11d7a331fa7b4122ac2e2f1