[roll] Roll fuchsia [diagnostics][hierarchy] Bring back payload:null and align a few cases
Turns out we had an optimization for feedback snapshots in which we
don't include hierarchies with `payload: null`. Therefore I think we
should prefer that over `payload: { root: {} }` when we fully filter a
hierarchy.

This change brings alignment for that situation.

The tests reflect the cases that I consider to not be aligned before.

Run-All-Tests: true

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788324
Original-Revision: a5c812723b6b6b789c951e45075077fa61093ab3
GitOrigin-RevId: 60cd8bab57421f58536ef086aea25af87e17cb00
Change-Id: I708a6ec2dbbc6e3b22d2a21944c3baa1f8d233df