[roll] Roll fuchsia [emu] Remove fx emu-remote
The `fx emu-remote` command has been broken for a while now, since it
had a dependency on `fx emu`. Even before that, it was going unused
simply because there's no demand for it; running over CRD works
perfectly well for all known use cases.

This CL removes the command, a test script that was meant to exercise
it, and removes a few references to it. There was also a test script for
`fx emu` that was previously missed that is included in this CL as well.

Original-Bug: 116978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788424
Original-Revision: 62ce539a68099575ba111113310aa80bcfe6d6c7
GitOrigin-RevId: 330805df0c9c5d6b5aed0f1c86e64b80ce021598
Change-Id: Ia10428dc22ca4cf2924bbec6e037854e2bec4a99