[roll] Roll fuchsia [netstack3] Serialize TCP segments lazily
Add a method to BufferProvider that allocates without reuse and use it
to implement Serializer for SendPayload. The new method is necessary
because SendPayload is only a view and doesn't hold a buffer that can be
reused. Implementing Serializer for SendPayload allows it to be passed
in as the inner body of a packet.

Previously, the contents of the SendPayload were being copied into a
Vec, which was then being encapsulated in other packet builders. This
resulted in a double allocation: once for SendPayload -> Vec, then again
to create a new Vec to hold the entire frame/packet. This change removes
the first allocation by deferring copying from the SendPayload until the
final Vec is allocated for the entire serializer.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786703
Original-Revision: 069dec91c66ae13ef58ddedb49377b8f57472052
GitOrigin-RevId: 7a1abca12451333885ca26c9bd27023ae5527c87
Change-Id: Iead7afa93379e850bbedc010236b53d47f073f28