[roll] Roll fuchsia [miscsvc][assembly] Migrate to a package
zxcrypt was updated to look for the config file in two locations - /pkg
and /boot. This is necessary because when miscsvc is packaged it no
longer is presented with /boot as if it were /pkg. In order to vary the
contents of it's /pkg per board, we would need to make it into a board
specific component. While switching to structured config would allow us
to remedy that, it is outside the scope of this change to switch miscsvc
over to structured config.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/784125
Original-Revision: 855183c2098eec53997eb5aca1fba59499df5b77
GitOrigin-RevId: eb32706e1863e04678c8f39ff32beffaf0beb669
Change-Id: Ia4d78470daa4856ed078abc3ddce5035b6b06e58