[roll] Roll fuchsia [kernel][phys] Record phys Log in memory
This adds machinery in physboot to make stdout a tee between a
resizable buffer of contiguous whole pages and the console.
Everything written to stdout is mirrored to the log buffer.
debugf always writes to the log, but still only writes to the
console with kernel.phys.verbose=true.

For now, the buffer isn't used, just grown in memory and then
discarded by handoff.  A subsequent change will use it for
instrumentation symbolization and handoff.

Original-Bug: 95625
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/784702
Original-Revision: ac668931c8ff277def6cc568879bc3b008f766a1
GitOrigin-RevId: d57d7b36339c5757be2b47429d2331906a722461
Change-Id: I02f492f3a800311710e24a413e50c1cc3b3449e4