[roll] Roll fuchsia [network] Add `event_stream_from_state` helpers
This method instantiates either a `WatcherV4` or a `WatcherV6` client,
and returns events from the watcher as a `Stream`. This abstracts over
the Hanging-Get style API.

Additionally this CL introduces fake implementations of the `StateV4`,
`StateV6`, `WatcherV4` and `WatcherV6` protocols as part of a `testutil`
mod. This was necessary to test `event_stream_from_state` and may prove
useful to future clients.

Original-Bug: 115870
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762867
Original-Revision: 73ba939cb414a06e62f6701684af20d09829d6dd
GitOrigin-RevId: 9835be1fda5a357e9fc6ac1ac547bea5e0d520b4
Change-Id: I932d0a25632c299880b431e7eac1912e4f01694c