[roll] Roll fuchsia [build][bazel] Make the Bazel TOPDIR more configurable
The location of the Bazel top directory, relative to the Ninja output
dir is now officially determined by the value store in a configuration
file at `//build/bazel/config/main_workspace_top_dir`, which is read
by GN and shell or python scripts alike.

This determines the location of the workspace/ and output_base/
directories for the Bazel workspace, and a few other things.

All other references to this sub-dir are removed (except in comments
and documentation).

This will allow moving the workspace to a new location in the future,
with a very simple CL that changes the value of that file (and
comments + documentation).

+ Add tools/devshell/lib/bazel_utils.sh to export the fx-bazel-top-dir
  function, then make 'fx bazel` and `fx get-workspace-dir` use it.

Original-Bug: 113116

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779045
Original-Revision: 00b697b9402f61f34cf138adf93da250e2c14d38
GitOrigin-RevId: 2d484448d3cbf580b7ba8e5ef75369ccf7fad145
Change-Id: I74c48f15a323363931fc50d13e6c0736c85d04c2