[roll] Roll fuchsia [dhcp/client] Use option code consts; drop domain name
Instead of hard-coding the magic numbers to denote the parameters
in the parameter request list, use the optionCode constant definitions.

Doing this uncovered that we were requesting optDomainName but not
actually consuming it at all, so we can stop requesting it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788563
Original-Revision: 3fc1cc1847df42583c0083e7ef3ff27c029744b9
GitOrigin-RevId: 42d7b1f86e05ee889d25784bfd6a3af8e087bc0a
Change-Id: Ieb964f88d364aaf89e7c1d0bc5b9e0a76dfd654b