[roll] Roll fuchsia [emu] Tweak path handling logic for --device
The attached bug describes how running `ffx emu start` with no --device
flag prints a message about using a path as a virtual device file, which
makes no sense to a user who didn't enter such a path.

This cl adjusts the handling of the Some("") case for the --device flag,
such that it is treated the same as None. It also short-circuits the
parsing logic when the path given is not a valid file.

Original-Bug: 119271
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788422
Original-Revision: d2b3834dffaf6c60cfbf399e6fcc68a33e244b89
GitOrigin-RevId: 3fc7e1aac2078d9cd4b27fcf962c7a5ada543a01
Change-Id: I49bc0478b3b0815ded2a644345d51a3242e72e0d