[roll] Roll fuchsia [doc-checker] Detecting brackets that are not links
This adds code to check for the situation where the
markdown contains [something]. This is ambiguous since
it could be an autolink to the reference [something],
or it could be simply text.

This change compares the contents of the link to the link_url
and if they are identical, it is flagged as an ignored autolink.

This generates an info message, but does not block the submission.

Note: This change is only visible if you use the
`--check-reference-links` flag.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788605
Original-Revision: 7b034cca7bf88b26bee21ed19cb3cbbabf62553a
GitOrigin-RevId: 05506fcba341343f31ede01dda32495feb90b2f0
Change-Id: I96080f558879e0be6e7ecb39c4a2860cb22c670d