[roll] Roll fuchsia [reachability] Ignore Stale neighbor state
Neighbor health can't be properly inferred from the Stale state. By
considering the stale state as possibly healthy, the watchdog missed a
case where the tx path was blocked but not rx. The gateway then can send
probes to the Fuchsia node, which move it to the stale state repeatedly
and cause the watchdog not to fire.

Ignoring the stale state improves predictability here and neighbor
health is only transitioned in the Unreachable, Incomplete, and
Reachable states.

Original-Fixed: 118055
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788840
Original-Revision: d92d696e86dbb3312936bda2ea3cd8e1a50da851
GitOrigin-RevId: 81648233f16f4164ce6d8eb63fb6ca2ab6771252
Change-Id: I14341b2aaa54b64cf844ccd83554550ba2d54228