[roll] Roll fuchsia [fidl][dynsuite] Server suite separate controller
Change the server suite to use a separate controller protocol instead of
the reporter protocol. In addition to sending reports on one-way methods
received by the server under test, the controller protocol is also used
by the harness to command the SUT to send events, and has the ability to
report the result of sending an event in addition to just commanding it
to be sent. This separation of the "send event" command from the SUT
also avoids some potential pitfalls of manually building the "send
event" command and sending it on the same channel as the SUT.

Original-Bug: 102339
Test: fx test //src/tests/fidl/server_suite
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786302
Original-Revision: ee80752f41c396a885a31128ecdb98419fd9af64
GitOrigin-RevId: 1ca0b211d56812e999e66cb6d2279f6b4e9e6020
Change-Id: I2d36c1db33cf467d32cdabdd5c8bdc9453731c56