[roll] Roll fuchsia [stress-tests][fxfs] Add progress timer
If no files are created after 60s, a timer will trip that will cause the
test to fail.  This helps us ensure that forward progress is made (which
should be guaranteed since we are also deleting files in another task).

60s is arbitrary; it's big enough that flakes are unlikely.  We could go
even higher if we wanted to, since these tests run for fairly long
periods of time.  The period at which the `crash` variant restarts the
test environment is increased to 90s so that the timer isn't completely
useless for that test.

Original-Fixed: 110593
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783882
Original-Revision: 48d6c55182bb7f034798ef439d93408691f86fe5
GitOrigin-RevId: 5d98fbc18007fab200f55a24a287475f307dfb55
Change-Id: If9c13d6eb05bfdbb25398fdef4175eab26d4f136