[roll] Roll fuchsia [pager][fxfs] Support zero range bit on writeback
Add support to the rust pager wrappers for passing a client's
expectation that a range is entirely zero when the client begins to
clean a range.

This is used to avoid a race where a page can be dirtied between
query_dirty_ranges returning that the page is all zeroes and calling
writeback_begin with the page.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785907
Original-Revision: 10dd17699fc0d31801022e3d5374c12a94a3ce83
GitOrigin-RevId: 1c69fe471c437c22e9e92dc306890b76e91348dc
Change-Id: I3e08087f8911807ea4705478990b04dfbae22f49