[roll] Roll fuchsia [component][fuzzing] Check for message limits in ChildProcess
This CL adds checks to `ChildProcess` methods such as `AddArg`,
`SetEnvVar`, and `AddChannel` to ensure the parameters passed to
`fdio_spawn` can fit in a FIDL message.

Previously, the easiest way to exceed these limits and thereby fail to
spawn a process was to provide a libFuzzer-based binary with a large
number of individual inputs to try. When run as a unit test, the engine
tries all inputs in the seed corpus. Simply providing a sufficiently
large seed corpus would cause tests to fail.

As a result, the LibFuzzerRunner has also been updated to use the
feedback from its ChildProcess object to batch inputs into several tries
when they would not fit into one.

As a result of testing, a bug was observed and corrected where the
fuzzer tests were not converting relative paths to be package-absolute.

Finally, in order to adequately test all of this, the sample corpus
under common/testing/data was dramatically increased to include 2000
separate inputs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788428
Original-Revision: ccef0da45e4d36750c20897a7de3cf488a2fa84c
GitOrigin-RevId: d8858118c35ba4dbc175eff8569e428f2dbd9b34
Change-Id: If044dee18e8e28507f5fe80fcdae9853b4b5d511