[roll] Roll fuchsia [settings] Box pending_requests in setting proxy
Heap profiling has shown that the setting proxy is responsible for a
fair amount of memory. One way to trim that down is to reduce the size
of our `ProxyRequest` enum by `Box`ing the `RequestInfo` struct. After
some refactoring we can see this suggestion from clippy:

warning: large size difference between variants
   --> ../../src/settings/service/src/handler/setting_proxy.rs:110:1
    |
110 | / enum ProxyRequest {
111 | |     /// Adds a request to the pending request queue.
112 | |     Add(RequestInfo),
    | |     ---------------- the largest variant contains at least 376 bytes
113 | |     /// Executes the next pending request, recreating the handler if the
...   |
117 | |     HandleResult(SettingHandlerResult),
    | |     ---------------------------------- the second-largest variant contains at least 136 bytes
...   |
130 | |     Listen(ListenEvent),
131 | | }
    | |_^ the entire enum is at least 376 bytes
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#large_enum_variant
    = note: `-D clippy::large-enum-variant` implied by `-D warnings`
help: consider boxing the large fields to reduce the total size of the enum
    |
112 |     Add(Box<RequestInfo>),
    |         ~~~~~~~~~~~~~~~~

Since proxy requests are always added to `pending_requests` we also
switch to having that hold a `Box` as well to avoid an additional copy.
We could consider doing this for `listen_requests` as well for a follow
up but the way things are structured we'd still need a copy so that's
less of a win.

Tests: fx setui_service_tests
Original-Bug: 119455

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789081
Original-Revision: 94eada301f09b9e3670adf82ee65e6942055512c
GitOrigin-RevId: 78584bf04d98482a142a84f81a212df2ebe4a93c
Change-Id: I9bf3fc7725e406dcd1835c1ae143fb23ababd1a0