[roll] Roll fuchsia [starnix] Handle ptmx not being in /dev/pts
Depending on the distribution, the ptmx file can either be a symlink to
/dev/pts/ptmx, or another device node. Because it must access /dev/pts,
instead of relying on the node being on the right file system, this CL
pass the root node from the /dev/pts filesystem to the node operations
on ptmx. This is tested by BasicPtyTest.ReplicaEntryGoneAfterMasterClose
when ptmx is a device on /dev instead of a symbolic link.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788336
Original-Revision: dafb3a976623f348283bec6b88880ba39cfe812d
GitOrigin-RevId: ad5dc9e87750a570898e3d344895de41306bb0ca
Change-Id: I8569b1a2ed0442309ad8287563714af11d1395c4