[roll] Roll fuchsia [driver_runtime] Rename fdf_arena_destroy.
This will make it more clear that your are destroying a reference
to the arena. If we do want to change the ownership semantics
in future, it shouldn't be too hard to rename as it's rare
anyone would call it directly.

Original-Bug: 114821

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789506
Original-Revision: f545f99aa79ba70dddcbd1b6bace2d5496be3041
GitOrigin-RevId: 05ac09e91b7f6664b6b0a1216c364b1d91b14cfb
Change-Id: I4ffc8c39086a9e4fa9212189acad80cc0ca2e1f6