[roll] Roll fuchsia [usb][dwc2] Set clear-NAK and ep-enable bits on timeout
This sets the IN/OUT EP0 clear-NAK and ep-enable bits in preparation of
the tokens to follow a timeout interrupt. The tokens to follow will
either be a SETUP token if the host received the data but its ACK was
lost, or an IN token if the data sent to the host was lost in transit.

Original-Bug: 105382
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789272
Original-Revision: 32fdd85dd230fa2bd5e7e0b70b8eabd06a801e81
GitOrigin-RevId: 2175f30185552be5e95e7745adfb0bd35e6c7495
Change-Id: I28ebe520242ec698fd835a8ff4ba676cb8b1bd2a