[roll] Roll fuchsia [emu] Swap Mutex guards for serial tests
The attached bug explains how we recently discovered that using automock
on async functions is unstable, and attempting to work around this with
a Mutex is insufficient since the mutex is thread-owned and thus does
not guard the static mock contexts between the various tests.

This CL removes the mutex and replaces it with annotations to serial the
tests that utilize the mocking context instead. By serializing the
tests, we give them atomicity and ensure that the context is set, used,
and destroyed all before the next test starts. While this can slow down
the test execution, these are very fast tests and the delay is not
enough to notice or care.

Original-Bug: 119219
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789178
Original-Revision: 0e05abe95c0ed8213fa52eef6ffb7eff6e8e2e0b
GitOrigin-RevId: 1a7e103c504a98046a4ad9836030d3f387713e56
Change-Id: I734eff44ce4074fecfce5e941dcf022cbc9d059f