[roll] Roll fuchsia [netstack3] Put Keep-Alive and Nagle options into a single struct
This gives us the benefit of not adding more and more accessors for each
of the socket options that we add in the future. This change also
removes the `NagleAlgorithm` enum and uses a boolean as a named field.
It does not worsen readability and aligns with existing mechanism used in
enabling keep-alive.

Original-Bug: 117115
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789262
Original-Revision: e5ee7337889242ba604093d1789572bf2cae6a8f
GitOrigin-RevId: abcab0c4d1142b5f2d3697774efbed0d39c8b972
Change-Id: Ib49e5c5da1497e1a5fa3c4bded006d40e3b15048