[roll] Roll fuchsia [tefmocheck] Suspend test may cause an OOPS while passing.
The suspend test checks that device successfully reaches the S3 (Suspend
To Ram) power state and then attempts to resume. At this stage we do not
expect resume to succeed and in fact it is likely there will be multiple
OOPSes caused by leaking BTIs. This does not reflect a failure of the
test, but simply the expected state of resume currently.

Original-Bug: 117528
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788273
Original-Revision: 2ca60316114043527964a62424180140d9101514
GitOrigin-RevId: 6210e588863241e66d28a4892aa0d4d428b3cc00
Change-Id: If17676b86f8f112da9032053eef149c513b2957c