[roll] Roll fuchsia [assembly] Rename EmptyMinFS -> EmptyData
In the future we need to support both minfs and fxfs. This makes it more
clear that the empty data partition may be formatted with either minfs
or fxfs on some products.

Add TODOs for eventually removing the --with-empty-minfs flag.

Test: No functional change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789123
Original-Revision: 492d8d365a3a66f3871f0db1f4493256add1a853
GitOrigin-RevId: b33202ec8483e6aab5da9b28745eff50cfa1dad8
Change-Id: Ib9f2329f01973c076df07ea9489f18944080a6b2