[roll] Roll fuchsia [ffx][error] Downcast anyhow errors, not type check
When I changed over to the new error type I mistakenly thought doing a
`.is<T>` check on the anyhow error would give me a result in line with
the likely success of `.downcast<T>`, but it doesn't. Downcasting
actually *searches* the error for a matching type of error, so it'll
pull one up from deeper in the error stack.

This fixes the logic to use downcast instead, resulting in it reaching
down the error stack to find an ffx user error.

Original-Fixed: 119301
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789956
Original-Revision: d74a2ff4ff7313be969a7bcf9fce8b7147306e8e
GitOrigin-RevId: 145057144d1b220aaa7481989f4894e1b2c89beb
Change-Id: I22370eb3ed6ed41059827b036227756dc829bd35