[roll] Roll fuchsia [emu] Add testing around control sequence of start
This CL adds more tests to the emulator start subcommand, specifically
in the domain of the start command's control sequence. It verifies that
the dry-run, stage, and edit flags interact correctly with the config,
reuse, and verbose flags and that changes in each stage are correctly
propagated to the final start Command.

In writing these tests, we uncovered an as-yet undetected bug in the
--edit flag behavior that was actually losing the user's modifications
before launching the emulator instance. That has been addressed in this
CL as well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789729
Original-Revision: 198463783e21fa862ef2fbb04b0c2bdaa31faa54
GitOrigin-RevId: 16362c2e7da7cc073e0e1d3ce9dee9f5fffb9f37
Change-Id: I94ec7e570e753ac9ab3040a946cc8f99a5ca0022