[roll] Roll fuchsia [fuchsia.driver.framework] Remove available attribute
This attribute is unused because this FIDL is in the experimental
SDK, it is not in the partner SDK yet.

This may have been incorrectly added in fxr/664072

This available annotation is not true at the moment, and it prevents
us from easily making changes to these definitions.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790482
Original-Revision: 0dd315d73dad9887bbc3aeef10c9c85c1673cdeb
GitOrigin-RevId: 1056e14ff85b696f087fa8bca504efd8ef8fc88c
Change-Id: I79a6ccca4b812bdf160c5fa2c1da6eb99e6c9fce