[roll] Roll fuchsia [archivist] Gracefully handle invalid UTF8 in log messages
Instead of returning an error when getting invalid UTF8, we can instead
replace the invalid UTF8 sequences with �.

The Archivist will no longer close log sockets when it fails to parse a
message received through the socket, instead it will leverage the
existing logic handling errors in StoredMessage.parse to return a Data
with an error set explaining that the message couldn't be parsed. When
invalid UTF-8 is found, no error will be set, but the message will
contain the � for each of the invalid UTF-8 sequences.

Original-Fixed: 119194

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790122
Original-Revision: efe5c14140f07b90f19d80c4ab50b4ec6abfd3cd
GitOrigin-RevId: 1cce5db2b86a1fe3141176a5c1d4e9861a60551b
Change-Id: Iaa495e42ed02db661446b3f919eb77ff21d24f1a