[roll] Roll fuchsia [realm-builder] update docs and non-API-breaking idents for fragment-only URLs
For historical reasons, the RealmBuilder API uses the term
"relative URL" when referring _only_ to fragment-only URLs, but Component
Framework (including RealmBuilder) now supports a new form of
"relative URL": relative path URLs that refer to components in a
subpackage.

This CL updates RealmBuilder documentation (particularly in code docs in
FIDL and in the multiple language-specific client APIs) to be more
accurate when referring to the different use cases for fragment-only
URLs vs. subpackaged component URLs.

It also updates internal identifiers (as long as it is not API-breaking)
to use a more precise name when referring to values that only support
fragment-only URLs.

Original-Bug: 118194
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787952
Original-Revision: b58103f355aa151477dc4032c8da7d8264f88a73
GitOrigin-RevId: c9f38eb60a2d7bb4c99f72bef7e2599af10d4534
Change-Id: I4db9291846e20e505b1a8c32dcaffe4f3fe3c1b2