[roll] Roll fuchsia [ffx][metrics] Abstract metrics logic from ffx main
This moves the logic for metrics reporting out of the main ffx run
function, allowing it to be run by subtool processors and putting the
argument redaction logic in the right place. This change is intended to
be a code shuffle with no major logic changes, follow up changes will
actually do the work.

There is one logic change though:
This *does* make it so that when run in isolate mode, metrics
will be turned off.

Original-Bug: 116489
Original-Fixed: 114678
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789517
Original-Revision: 5fbd52e1399257960f458d910e222fb630539d43
GitOrigin-RevId: 81dfa6d9eed3ebd1ee344e98ed64b1488d752411
Change-Id: I04cef35c8070d867c30ff95ca891eb518c37ca07