[roll] Roll fuchsia [input] Don't block input pipeline if reader stuck
When an event is handled and sent to registered listeners, it is possible a listener's reader might get stuck and never send a response. Because of this, we want `handle_input_event` to return after sending the event and we want to make sure other listeners can still receive the event if a reader gets stuck and that subsequent events can still be handled.

Original-Fixed: 106843
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/771343
Original-Revision: 1eef17b5999dd95a47b7af43b147a1976ce7677d
GitOrigin-RevId: 2996e1e51195e7bc8b1a1c03818bee4fe9e40a29
Change-Id: I48a07930177b573ef75c2f222a526ea5ff989dc9