[roll] Roll fuchsia [debugger] Improve the UI while loading lots of symbols
When loading and indexing very large symbol files (chromium, for
example), the user is just stuck with a non-responsive [zxdb] prompt,
which is not ideal.

This change aims to make that a little nicer by showing a "Loading
symbols for <process name>..." prompt that continuously shows the user
that *anything* is happening. Since there's really not a ton that can be
done before symbols are loaded, it's better to just not even present a
prompt than to move the indexing work to a background thread and show
them a responsive console.

FIXED: 6311
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783302
Original-Revision: 2ddd1f80bf7985df65be3aff79fa0a03eae87958
GitOrigin-RevId: 8aabd0c7c120e6f0129c0a0fc113383757ed4757
Change-Id: I1a4486bcaa58a9a373d1e590be2b4f850a41d5f2