[roll] Roll fuchsia [reachability] Adding integration tests using FIDL
Create a reachability component that incorporates the fake clock library
so that the test environment can be accelerated. The named_timeout
allows the code to pause prior to hitting the timeout window. Prior to
timing out, the code is accelerated, then slowed down to real time.

The integration tests cases where the internet_available flag was
initially true or initially false. This helps simulate the cases where
the internet went from good to bad or bad to good.

Additionally, testing of edge cases, such as the method of interacting
with the API, are included.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779254
Original-Revision: 9bc43c2ba2d5a831a6273a73db604845069e63a6
GitOrigin-RevId: 0cfaebe9898eea3806dd891d717ac0f4e4d5b735
Change-Id: Ib2d011eb8fd1d98ada5699c445b008090c3528e4