[roll] Roll infra/recipes [testing_requests] Simplify significantly
As a result of the deletion of the zbi_test.py and the merging of
botanist into testrunner, a lot of the complexity of this module is no
longer necessary.

For example:
- `api.testing_requests.task_requests()` can now take a test spec as a
  parameter instead of taking all separate parameters because the module
  is no longer used from recipes that don't use specs.
- The module now only needs a single entrypoint because it's only ever
  used by one production recipe.
- The `TaskRequester` abstraction is no longer useful.

This is a pure refactoring with no behavior changes, hence no generated
file changes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/790602
Original-Revision: b1b2b4ec7ec809c5b45e5beae51939994d6d52bd
GitOrigin-RevId: 7b6c9eb3f804884145286abae18dd4d105fcad21
Change-Id: I03ff100c3a37c46c84c3071bb2ef0a0e7fdc1cfb