[roll] Roll fuchsia [starnix] Pass galaxy outgoing dir to kernel
This change updates starnix_manager to pass the outgoing directory for
dynamic galaxies to the starnix_kernel.

This will allow the kernel to serve a component runner on behalf of the
galaxy.

This also removes the code passing the component controller from
starnix_manager to the kernel, since a future change will update the
manager to serve the component controller.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790922
Original-Revision: fe439074405af851f9981dcd433effa88feffa0e
GitOrigin-RevId: 92fb074a74e0660ccb2af48f0ca6307bb4998e24
Change-Id: Ieac576a42cf0f31f6ba75544f890d39437f477ad