[roll] Roll fuchsia [netstack3] Accept ZonedAddr for TCP bind
Use the zone information in the address passed to a TCP socket bind()
call if it is present. Implement this the same way as for UDP, where
binding to an address with a zone implicitly sets the bound device for
the socket.

Original-Bug: 115524
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788334
Original-Revision: 891a865a04d351b2344e7332b4ac6b0e127bf750
GitOrigin-RevId: b01df69415786fdf4d9cc41f850821313931288f
Change-Id: I23f3b36090ba2eb20968aedeba77c989cb549a5a