[roll] Roll fuchsia [xhci] fix HubDeviceRemoved race condition
If HubDeviceRemoved was called during NormalRequestQueue when the lock
was released for the WaitForIsochronousReady call, the TR would have
allocated a TRB context already, but when HubDeviceRemoved completes, it
would've disconnected the device and dealt with all TRBs in
pending_trbs_. After the lock is released in HubDeviceRemoved,
NormalRequestQueue may continue running, but it doesn't check again if
the device is disconnected or not and commits the TRB, placing it into
pending_trbs_. This TRB context does not get cleaned up and causes a
crash when deallocating.

This is fixed by adding checks after we re-acquire the lock after the
WaitForIsochronousReady call.

Original-Bug: 119477

Test: tested on Atlas

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790511
Original-Revision: 9ce3b3b08cb5b2ab0a48e89cdbc7cf184f78c860
GitOrigin-RevId: 66e2a7159830bb2fa0adab1c83ba81e7b986ba34
Change-Id: I3657414ecc1414a2be433caa96ccf073b3b2d5de