[roll] Roll fuchsia [flatland] Remove locks around VkRenderer::ExtractImage()
VkRenderer::ExtractImage() is a potentially expensive method
called from a worker thread. This CL adds threading annotations
to ensure it is never locked, and removes the current locks around
it.

Also found a spot where we make an unnecessary call to sysmem (we
were checking that a buffer was allocated in ImportReadbackImage()
even though we'd already checked the same thing before the call) and
removed it.

Original-Fixed: 119195
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790518
Original-Revision: 967c142aba001dd9411aeb95c5f85c6e6156a8cd
GitOrigin-RevId: 601696ec113a2046b122181648b71add153c8fe4
Change-Id: Ibbc6c11c758c7d538aedfe54353b30963b3c1072