[roll] Roll fuchsia [tools][sdk-tools]: adds functions to SDKProperties to accept a context
Adding a context.Context to common SDK commands that may be long running
(e.g. ssh, gsutil, ffx) to allow for cleaner process cancellation and
cleanup.

This also allows for a logger to be passed through the Context (to be done
in a later commit).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788310
Original-Revision: 6dfa7123b2b733f8a40f9fe8424340803ec95fee
GitOrigin-RevId: 808b758bad62d11f90f2cf74ec25e7a60cecc16a
Change-Id: Ie7329235ad308028f4575e18149aa5bf32253421