[roll] Roll fuchsia Reland "[starnix] Use json5 test expectations for most tests"
This reverts commit 937e2995e5ce5495bb4bfe539cd49ddf13e9820b.

Reason for revert: Did not fix flakes.

Also removes expectations for libutils_test.
LooperTest.MultiThreaded_NoUnexpectedFdRemoval flakes when run by itself
because the test case races with the test trying to connect to logd. The
test creates a socket when trying to connect to logd - if that socket is
created between the two pipe() calls in the test, the socket will be
assigned the fd that the test expects pipe to receive. This doesn't
happen when all the test cases are run together because the socket is
created before the test case is reached.

Original-Bug: 119660

Original change's description:
> Revert "[starnix] Use json5 test expectations for most tests"
>
> This reverts commit e5f3df491b80448fbf34d22745fad8ddb9fd9059.
>
> Reason for revert: Seems to have introduced a lot of flakes.
>
> Original change's description:
> > [starnix] Use json5 test expectations for most tests
> >
> > This updates most gtest/gunit based test cases that have expectations in
> > the Starnix tree to use the json5 test expectations scheme. The one
> > holdout is the in-tree fdio_test which uses a different output format
> > and supports a different set of parameters due to using a different
> > version of gtest.
> >
> > Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788432
>
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790042

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790462
Original-Revision: 65be9ec85c0c4ed872ebe80ae27e32ef47e59b0a
GitOrigin-RevId: 7ce9eee22e6fe452ace25ffdfeaa5981df80c9f2
Change-Id: I964882bd3912d9af518a5947bd5c6cb70ed35c3b