[roll] Roll fuchsia [ffx] Re-organize how FfxCommandLine is used
This cleans up the kind of confusing path the ffx command line takes at
program start by combining the Ffx args struct into the FfxCommandLine
struct that holds on to the original invocation, and replaces most uses
of one or the other with the new combined struct.

This also removes it from the ToolSuite from_env, which never really
made much sense. You shouldn't really need a parsed command line to
create the tool suite, so long as you have an initialized environment.
Nothing ever really did anything with it this early anyways.

This also unifies all help command listing output around our own version
of write_description, which will be helpful in the future, and breaks up
the appending of the command list into its own thing to get rid of a
turbofish.

This is mostly just cleanup to allow the next change to be possible, but
it was long overdue cleanup.

Original-Bug: 116489
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789518
Original-Revision: 0ddc56b1f1f6d47ed6bc6745e7a579590820beb7
GitOrigin-RevId: 3333901279694343f98496d466c6cb96e21cd96d
Change-Id: If9fa1ff78361695e1b996a47338a9c9e3de3cacb