[roll] Roll fuchsia [network/tests] Run NS2 tests with expectations
Rather than running all test targets with expectations, I've elected to
run only those targets for which we need to skip/fail a test. The
benefit is faster runtime and less configuration overhead; the cost is
that we support both formats of running tests. If this support becomes
cumbersome, we can elect to migrate the other targets later.

Original-Bug: 115159
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790663
Original-Revision: 31ccc0781d403f0cd0268753077ee2debceb2892
GitOrigin-RevId: da6a471737e68fadc947b43bb1678a90b3644d03
Change-Id: Ia5c6b04f42a4cc4c98afee862303d7ba011c6845