[roll] Roll fuchsia [iwlwifi] Create a new dispatcher for FIDL server in wlanphy-impl-device
This CL partially reverts fxr/761123.

wlan-softmac-device and wlanphy-impl-device are in the same driver host
and share the same default driver dispatcher, the device unbind and
remove tasks are all scheduled on this dispatcher. Thus, the
sync_completion_wait in CreateIface() in wlanphy-impl-device will block
the tasks on this dispatcher including unbind and remove tasks. So no
matter how long we wait in CreateIface(), the device unbind/remove
tasks will wait for the completion of CreateIface(), and CreateIface()
always completes as a timeout since it waits for the tasks that wait
for it.

This change creates a separate dispatcher to do the job.

Original-Bug: 118079
Test: Delete iface 0 with "wlan-dev iface del 0" and no interference
observed. New iface is created successfully.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781589
Original-Revision: 7977906e67c57faa79c7b4d89e7fc648b5336f55
GitOrigin-RevId: fceced3e8552ded424ff0965e87dc793dcbfd5a4
Change-Id: Ie94b099b27bef706866be40f6c6c8552899c6663