[roll] Roll fuchsia Reland "[fdf] Skip msd integration test in DFv2 builder"
This is a reland of commit ede524e68e8cb261eb87f236be33e7a831b8c857

This CL was reverted because it resulted in buildbot
workstation_eng.chromebook-x64-dfv2-release failing to build. This is
because the buildbot references a target in
//bundles/buildbot/dfv2/BUILD.gn which GN failed to process.  GN failed
to process the build file because the target
//bundles/buildbot/dfv2:core failed to be created. One of the
dependencies of //bundles/builderbot/dfv2:core introduced in this
original CL was updated after the CL was merged resulting in that target
failing to be created. //bundles/buildbot/dfv2:core has been moved to
its own file in order to avoid breaking the DF2 workstation buildbot and
the dependency has been fixed so that //bundles/buildbot/dfv2/core can
be created.

Original change's description:
> [fdf] Skip msd integration test in DFv2 builder
>
> Flatten some of the dependencies of //bundles/builderbot/dfv2:core so
> that
> //src/graphics/drivers/msd-intel-gen/tests:msd_intel_gen_integration_tests
> can be removed from being a dependency of
> //bundles/builderbot/dfv2:core. This is so the dfv2 core builder bot
> will not run the test. It is currently failing when dfv2 is enabled.
> fxbug.dev/114096 tracks fixing this test.
>
> Original-Bug: 102234
> Original-Bug: 114096
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776168

Original-Bug: 102234
Original-Bug: 114096
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790562
Original-Revision: 15409165714e7d8e4d49525f49d2a2881d97b204
GitOrigin-RevId: 3a274b67ed423a2b084d0d1c2334a0c3eb1f4df4
Change-Id: I7fd79cd642fc175375ae970c5edf34a02121fdfe