[roll] Roll fuchsia [iwlwifi] Reset wait_for_delete completion after signaling it
"wait_for_delete" completion is used by phy_create_iface() for waiting
for in-progress iface deletion. When an iface finishes the deletion,
"wait_for_delete" will be signaled, but since it's not reset, it won't
be ready to be waited on for the next phy_create_iface(). This change
resets this completion after signaling it.

Original-Bug: 118079
Test: Manually deleted iface for 5 times with "wlan-dev iface del [id]",
and all of them got recovered successfully.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783895
Original-Revision: cca4de5cb607355fd4335cbc301f07245a96bb04
GitOrigin-RevId: 69fe1c7de3d5e50c5c3b1b90d8a134e748bd7208
Change-Id: I239c8cec3d039f7a4945d8324dd7c918e7f2c4ac