[roll] Roll fuchsia [components] adjust regex used in testing URL validation
The URL validation portion of the unit tests for the cm_fidl_validator
library uses proptest to generate random input to the validator based on
a regex. This regex allowed invalid URLs to be generated, and thus needs
to be adjusted to prevent this scenario.

The new regex is constructed from the ABNF specification in the URI
standard (RFC 3986), but tailored to a subset of value that pass
URL constraints enforced by the `url::Url::parse()` parser.

Note that the special case "Invalid scheme" error is not actually
invalid, if the URL is valid as a relative URL, which is often the case,
so this error check has been removed and test cases updated. (Including
this extra error check would break proptest!() because the regular
expression for proptest!() could generate a legal URL that still failed
the scheme test.

MULTIPLY: cm_fidl_validator_url_proptests
Original-Fixed: 118872
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785852
Original-Revision: d35f57c45c838d475032c5ebdd05dc3b341404c1
GitOrigin-RevId: 6e1cb028b0602d65a11831dbbf0118b51dc822f2
Change-Id: I950cc86cfb26e978c8fb97aca75fd47ca5f14822