[roll] Roll fuchsia [fuchsia.io] Node2.Reopen takes a server_end:Node
Node2.Reopen should only be able to create connections to Nodes, so
the flexibility of zx.handle:CHANNEL is not needed and the type can
be narrowed to server_end:Node.

This is a breaking change; it cannot be expressed using platform
versioning because to do so would require conditional compilation
support in Dart (see https://fxbug.dev/105932).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790825
Original-Revision: 3de15a72ea5a9cb0d2825681c385ea2620eb9a28
GitOrigin-RevId: 99b958a3c3d8afb5bf5e4d6614b5e9ca26036e2b
Change-Id: I0019acd8c8a07976dfdbb7207b1da7ff75f1598f