[roll] Roll fuchsia [elf_runner] Use smaller buffers to forward stdio to logs.
Limits the size of the forwarding buffer to the number of bytes
which the kernel says a socket has to read, up to the maximum
message size that we'll forward to logs. While at times this may
grow to as much as 50% of the size of the previous implementations'
buffers, we also use fresh heap allocations for the internal buffer
each time a message is yielded so that large allocations will be
quickly freed after being forwarded to the logs.

Adds a new unit test which replicates the behavior expected by
some intl examples' integration tests.

Results in a 500-800KiB reduction in heap usage by Component
Manager in the product configurations I tested. While the previous
implementation allocated ~10MiB of buffers in some configurations,
most of the pages for those allocations were never touched and thus
never actually provisioned by the vmm. That said, the previous
implementation still had the potential to allocate all of those pages
in pathological scenarios, so in addition to the savings these
smaller buffers should also result in more predictable memory
usage.

Original-Fixed: 118020
Run-All-Tests: true

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782162
Original-Revision: 277c41e874e2c2c7b98a1705c5d338bf0d01947b
GitOrigin-RevId: 94509b4c2f06bd83fe3a0ea53f69f51cf96a3125
Change-Id: If5f8af0841375aff6f09fefdd0c07bdd43eb91df