[roll] Roll fuchsia [superproject] Roll cobalt Remove CHECK failure if FileObservationStore can't create a directory.
This is consistent with the other uses of MakeDirectory in the
ImmediateLocalAggregateStorage. All the other uses of the directory
and file system in the FileObservationStore seem to fail gracefully.
So this should only result in dropping observations, and not some
other failure downstream of the CHECK.

Original-Bug: b/265935870
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/791164
Original-Revision: 44b782c5734825fab4ad267a59c49860cc9e6362
GitOrigin-RevId: 582f132bc7a62fb3c68f0ae4cfeb2f7c650a53f1
Roller-URL: https://ci.chromium.org/b/8791494099934573345
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791810
Original-Revision: 2f6b5ba716fc17236bffbf21061faf837680c059
Change-Id: I883027d64a2dba3292d332e0fdd448e7e8a7e3ea