[roll] Roll fuchsia [build] Remove todo regarding investigating asan odr indicator
This has been investigated. The asan odr checkecing requires
default visibility symbols for the odr indicators to check
if the indicators have been interposed upon. This implies that
a symbol of the same name was found at runtime.

It's that clear that we will ever want to turn
-fsanitize-address-use-odr-indicator on. Having multiple definitions
of a symbol across shared objects is well defined in ELF.

Original-Bug: 114037

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791348
Original-Revision: 21c3c8bb3332c8db5d4d512f5062e02ea5a1fc28
GitOrigin-RevId: 4970a090fcec62483172065bac1430a7c56bd456
Change-Id: Ic64782f543befa836ea4e71a2452bdc9231a95a0