[roll] Roll fuchsia [ui] Migrate ft_device_test to safer teardown
This CL migrates focaltest-test.cm to use a single-threaded
dispatcher loop. The motivation for this change is that
|component::OutgoingDirectory| is thread-unsafe and should
not reference a multi-threaded dispatcher. Doing so will
often lead to flakes. If the |component::OutgoingDirectory|
is being destroyed, _while_ there's an in-flight FIDL request
on the internal VFS object, there's a risk for use-after-free
failures.

The pattern used here, with |PerformBlockingWork| is safe and
compatible with multi-threaded code.

Original-Bug: 118385
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791471
Original-Revision: 6a38ec36b8d32a8c4f0f76661738e178db2c6d96
GitOrigin-RevId: 9e637fa693aad8fe38ba259289a452fdb1c3bb67
Change-Id: I458435c774cb7491e1c80b125e1c229634fe473e