[roll] Roll fuchsia [zxio][1/3] Allow closing handles without waiting
This change introduces a new transitional function to allow for
`zxio_close()` to accept a `should_wait` parameter.  When `should_wait`
is set to `false`, this allows remote connections to drop the underlying
channel instead of waiting for a response from the remote server. This
can be useful when the result and errors of a close are not needed and
provides a significant performance boost in these scenarios.

Original-Bug: 118942
Original-Bug: 102657

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785462
Original-Revision: 4093d2cd5bdb051c969c892d66435dceb18de6a0
GitOrigin-RevId: c015753806bb6a72a32b22e34e44d20892f72ff6
Change-Id: I7adcce3ff7dae26560f9d4f431c4054b5e767a82