[roll] Roll fuchsia [iwlwifi] CancelSync should wait if task is executing
It's possible for the debug assert in TaskInternal's destructor to fail
if the destructor is called while the task is currently running.

CancelSync() should always return the task back to idle whether that
means waiting for the task to complete, doing nothing, or removing the
task from the queue. This commit changes CancelSync() to wait if the
task is currently executing.

Original-Bug: 120054
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791662
Original-Revision: 0e80624abe75fe1ea099379e903da833729149a4
GitOrigin-RevId: b39cbab9c98f264f6e9e0e73122977e0769bfccd
Change-Id: I6463248182cdb947745c04681bdddc00a07dba68