[roll] Roll fuchsia [fxfs] Enable LTO
This gains back the performance which was lost when splitting Fxfs into
several crates.  See, for example, WalkDirectoryTreeWarm.

The cost is a ~45% slowdown in compilation time for fxfs (19s -> 33s on
my machine).  If this ends up being problematic for productivity, we
could further split up Fxfs into a binary variant used for tests (which
won't have LTO) and another used for benchmarks and production (which
would have LTO), but that doesn't seem necessary given the build time is
still reasonable.

Note that LTO is not applied to the unit-test suite, so building the
unit-test suite won't be made slower.

Original-Bug: 119533

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791112
Original-Revision: b9be4d6ee9f70ff724609c62b7f9db43c447c913
GitOrigin-RevId: 2f8b9593a2aca4d69a21e87a011fec3b744f7b95
Change-Id: Idaf633e8e623bf0a155a41263b03e165c5acf732