[roll] Roll infra/recipes [testing_requests] Upload files directly from build dir
Instead of copying all the dependencies of each shard into a temporary
directory and uploading that directory to CAS, upload files directly
from the build directory. This eliminates the overhead of copying the
shard dependencies, which cuts about 1m15s off the duration of the CQ
long pole builder (core.x64-asan).

It also means that the paths at which files are accessed on the testing
bot correspond directly to the paths at which those files are produced
in the build directory (e.g. botanist is now at `./host_x64/botanist`
instead of `./botanist`) which is helpful for clarity and consistency.

The one downside is that we now need to write the botanist config file
and the test manifest into the build directory before uploading, which
pollutes the build directory with files not produced by ninja. I don't
think this is an issue as long as we give the files shard-specific names
to avoid collisions, and clean them up after uploading. testsharder
already writes package repository files to the build directory and those
don't get cleaned up, so I don't expect these new files to cause any
issues.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/791027
Original-Revision: 40a36a7a72491ef09d5e5275a91456ba2bf582d4
GitOrigin-RevId: 3c11647c92e491838cd12521f0e586bff94e169d
Change-Id: Ib24005d4612b0614300a4325e4fe77c23de76547