[roll] Roll fuchsia [versioning] Expose the target API level to non Fuchsia C++ targets
This exposes the target API level value to the non Fuchsia C++
targets (e.g. the host tools). This remove the definition that existed
in the debug codebase to make it usable more globally (and it also fix
the fact that this define was tracking the wrong value).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791523
Original-Revision: 4050e39bfb5a6c1a729bbac8a0761504fed826ca
GitOrigin-RevId: f86cf7651bd0a03910274cfd8d6a73ef28520949
Change-Id: I32fb75b4a0b3462b8230055699bff47c48011892