[roll] Roll fuchsia [ffx] Don't include any ffx command in subtool re-run
For some reason this was trying to include part of the ffx command line
in the re-run command for a discovered subtool. This meant that if you
ran it with `fx ffx`, it would trim off just the `fx`, but not the `ffx`
and produce a nonsensical command line like `ffx-repository ffx
repository`.

I'm honestly not sure why this was here, and it doesn't make sense.
Removing it makes that work again.

Original-Fixed: 120315
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791957
Original-Revision: af4798e54539421ae69832601cbd6be4455bde0e
GitOrigin-RevId: 2cbcd0422fd59bbc5b2e31681c2a58e3993a43c2
Change-Id: I2340fdb9294fbcc1709b1d8a0083fd6d21713f19