[roll] Roll fuchsia [fx][ffx] Only warn on `ffx repository default` mutations
`ffx repository default get` will return the default repository that's set
up by the tools/devshell/contrib/ffx wrapper script, so it should be safe
for users to call this method. This changes the wrapper to only emit a
warning if the user tries to run a command that mutates state, which
wouldn't be visible to users of this wrapper script.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791847
Original-Revision: 2b157281ce390d718504554a2c8bac79f4e8ee75
GitOrigin-RevId: 754c2b006008abc650eb213c2e9d8bc8913909ad
Change-Id: I64a9792efb1e7bddeee2d34c216e6d5d9ca9249c