[roll] Roll fuchsia [sdk] Remove fuchsia.modular.session from the partner SDK
This library is currently in the partner SDK category but it's not
exported in the SDK because it's not listed in
`//sdk/fidl:sdk_fidl_targets`. We could just add it to the list to
fix this but the fact that it's not exported means that it's
currently not used by the partners.

Original-Bug: 120471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792365
Original-Revision: 55ccac133b320a4e46cd8324c54c52dd843ca5b5
GitOrigin-RevId: c9e4fcb587902b8ddba90b7bfb79446e17b90c95
Change-Id: I6f90ad7407df0a7fd8bad4fe84a1921e806f58dc