[roll] Roll fuchsia [sdk] Remove fuchsia.input.interaction from the partner SDK
This library is currently in the partner SDK category but it's not
exported in the SDK because it's not listed in
`//sdk/fidl:sdk_fidl_targets`. We could just add it to the list to
fix this but the fact that it's not exported means that it's
currently not used by the partners.

Original-Bug: 120471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792363
Original-Revision: 6f090d2ff78c3a0d02737c7d672a241654eb5b13
GitOrigin-RevId: 37396619d9849bd1bfeec834fc311328f4803224
Change-Id: I0319e23eec82a02d1cb13577fd98f87792132d9e