[roll] Roll fuchsia [sdk] Remove fuchsia.hardware.temperature from the partner SDK
This library is currently in the partner SDK category but it's not
exported in the SDK because it's not listed in
`//sdk/fidl:sdk_fidl_targets`. We could just add it to the list to
fix this but the fact that it's not exported means that it's
currently not used by the partners.

Original-Bug: 120471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792364
Original-Revision: df8cd7118bc5deabe137c538abe53ae390226cb5
GitOrigin-RevId: f394df7fbc07860f368455108b3e00b898edd93f
Change-Id: I83f5256197d463c0756d8fecbb78b5e037c0eebb