[roll] Roll fuchsia [sdk] Remove fuchsia.ui.lifecycle from the partner SDK
This library is currently in the partner SDK category but it's not
exported in the SDK because it's not listed in
`//sdk/fidl:sdk_fidl_targets`. We could just add it to the list to
fix this but the fact that it's not exported means that it's
currently not used by the partners.

Original-Bug: 120471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792366
Original-Revision: 765ebb073617aba8550ea7493d07d281238105c8
GitOrigin-RevId: 64cc48a5fe01ec5a16694b84e6381671ce1709f3
Change-Id: I8919dc09fca45469cefac56c462e89a038289ec0