[roll] Roll fuchsia [sdk] Remove fuchsia.hardware.power.sensor from the partner SDK
This library is currently in the partner SDK category but it's not
exported in the SDK because it's not listed in
`//sdk/fidl:sdk_fidl_targets`. We could just add it to the list to
fix this but the fact that it's not exported means that it's
currently not used by the partners.

Original-Bug: 120471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792362
Original-Revision: f8e6b9ab977e0be469d95cc02d23cf09b8b98102
GitOrigin-RevId: 9e450c3766b20905dbf3b29e7d74322a7170dd19
Change-Id: Icc4c01eff1ae046449906ff9db3c6035187e55cc