[roll] Roll fuchsia [ffx][errors] Actually *use* IntoExitCode properly
Somehow this implementation was requiring that the Ok() case implement
IntoExitCode, but wasn't actually using it. This gets it back to the way
it was intended to be.

This makes it so that exit codes actually make it through the subtool
executor.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792210
Original-Revision: 14780bb6281062084ab05adcc6f5ccec53daa6f2
GitOrigin-RevId: 2419a9507e06a31e7acc42875fa68b6dac2f6ece
Change-Id: Ic243b02d5bc2a794cb5376c5aefa01faccdd9346