[roll] Roll fuchsia [sysmem2] sysmem2 tests
This ports all the sysmem(1) tests in sysmem_tests.cc to sysmem2.  For
the sysmem2 tests we use natural types everywhere.

This CL intentionally minimizes diffs to the sysmem(1) tests.  We can
potentially apply some of the syntactic simplification to the sysmem(1)
tests if we think that's worthwhile enough, but for this CL the intent
is to ensure the diffs are concentrated on creating the sysmem2 tests.

Token conversion (v1/v2 conversion) tests will be in a later CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763285
Original-Revision: c700beca29f370f8229f374345c0a8cfc4972d3a
GitOrigin-RevId: 9e3e0df019997232f8150f83ac2faf1a0263cce9
Change-Id: I16baf3541836929eef65c0427b221e22f1db31b8