[roll] Roll fuchsia [sysmem2] Combine V1 and V2 token serving to avoid ConvertToken() etc.
This way, a sysmem token is a sysmem token.  A client can use the
client end of the channel as a V1 or V2 token.

Specifically, since the sysmem(1) BufferCollectionToken and sysmem2
BufferCollectionToken are logically quite similar, and have identical
statefullness semantics, and are sent between processes to convey
participation in a sysmem BufferCollection, sysmem can serve both the
V1 and V2 token on the same channel.  This avoids needless
ConvertToken() messages that would also be needlessly creating
replacement channels, and the complexity in the server to cleanly
substitue a new channel for an old one.

Due to name collisions, this involves creating and serving a
fuchsia.sysmem2.internal/CombinedBufferCollectionToken, which has all
the messages of the V1 and V2 tokens, but with all names de-conflicted.
This internal combined protocol definiton uses @selector attributes
with full specification of library, protocol, and message name.

While we could potentially have the client directly use a combined
token protocol, keeping the V1 and V2 token protocol definitions
separate allows each to remain simpler to use for clients, without
wading past definitions for different-version messages.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/773524
Original-Revision: 16e640763b271919feb85fedd39fa483575eb9da
GitOrigin-RevId: a5ef127c9c4a3a43aad3a9dac1f5d3587de4ec4d
Change-Id: Ie467d5f37d386c8de2b5bd7083b3f38e905d9b6c