[roll] Roll fuchsia [netstack3] Support non-Vec insertion target
For TCP, there can be at most one listening socket on an address, which
means that an address entry should hold its listening socket in an
Option. In order to support SO_REUSEADDR, the TCP listener state for an
address will need to support inserting a listening socket when there are
already bound sockets present. That means the Option for the listener
will need to be the target of an insertion, which means the &mut Vec
target is no longer general enough.

This CL cleans up what was admittedly a cut corner by putting the &mut
Vec behind a trait that allows only a single insertion. This also opens
up the door for non-Vec implementations, which will be needed for future
CLs.

Original-Bug: 120272
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792380
Original-Revision: 0988da3c1362fb9693066f6a291a02f5aded62d0
GitOrigin-RevId: 6c6f0c7061987be5f184771747c0e1a7c5d3589e
Change-Id: I9f1e79bad2313c10269fa83c205031dcc40f1b44