[roll] Roll fuchsia [sdk] Remove fuchsia.cobalt from the partner SDK
This library is currently in the partner SDK category but it's not
exported in the SDK because it's not listed in
`//sdk/fidl:sdk_fidl_targets`. We could just add it to the list to
fix this but the fact that it's not exported means that it's
currently not used by the partners.

Original-Bug: 120471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792622
Original-Revision: 9b7840a36e562709756f27a303bcb38353125c6a
GitOrigin-RevId: 9a96fa7ba0d7ad6f387a75283da1c1d70d13d779
Change-Id: I0277ddbd48d6e6a9b04078460e95853d5888c236