[roll] Roll fuchsia [sdk] Add //zircon/vdso/zx:zx to the partner API tracking list
This is a follow up on
https://fuchsia-review.googlesource.com/c/fuchsia/+/792367 , this
library is also already exported in the SDK and so this CL is a no-op.

`//sdk/fidl:sdk_fidl_targets` contains the list of all the FIDL
libraries that should be exported in the GN SDK. Some of them aren't
currently listed and only end up being exported because another library
depends on them. This CL explicitly list all the partner libraries and a
future one will add some enforcement rules to prevent regressions. This
CL is a functional noop, the content of the SDK is unaffected.

In order to confirm that this is a no-op you can:
- Run `fx set core.x64`
- Build the `sdk:merged` target without this CL, archive
  out/default/sdk/archive/merged.tar.gz in a temp directory.
- Patch this CL, then build `sdk:merged` again.
- Compare the generated archive with the one saved earlier. The total
  number of files in both archives should be the same, as well as the
  list of FIDL libraries in the fidl/ subdirectories of these SDKs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792642
Original-Revision: 192cdd18fba1fb4afc93c92623352c0a043744b9
GitOrigin-RevId: 783a34dd6bf8e9edb38d0a3450211132fe261f16
Change-Id: Iebdd5aa6b145c440229b03e3980a4665356c9167