[roll] Roll fuchsia [fshost] add more context to errors in write_data_file
Add more context calls in the write_data_file path, so it's easier to
narrow down where errors are coming from. Also check the contents of the
written file in the tests, add a test for writing the file to the root
of the partition, and fix that behavior in the rust fshost.

Original-Bug: 120377
Test: fx test fshost_integration_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792262
Original-Revision: b9aab5e460764b4f32bf8e615d9c131d7de7cbc6
GitOrigin-RevId: 50da7c457daebb2b20b7b83753d87256fcab5002
Change-Id: I5e6dcaadfbc6b01d16daddd66a65530472c58845