[roll] Roll fuchsia [bt][fastpair] Report component metrics to Cobalt
Offer the Cobalt service from Cobalt to the Bluetooth-core realm.
Because bt-core is currently static, this definition exists in core.cml.
This will be removed when fxbug.dev/94055 is addressed and the
bluetooth-core realm is product-variable.

Route the Cobalt service to the Fast Pair component. This service is
marked as optional because it is not critical to component success. The
Fast Pair component is resilient to the unavailability of the
capability.

Log metrics occurrences any time a Fast Pair pairing completes
successfully or unsuccessfully.

Log metrics occurrences any time a Fast Pair request is received from
the Seeker.

Log the total Fast Pair pairing time to Cobalt.

File fxbug.dev/119634 to track logging Device Action requests to Cobalt.
Currently, Device Actions are considered Errors during parsing. Instead,
we should extract the relevant DA fields, log it, and record to metrics.

Original-Bug: 114083
Test: fx test bt-fastpair-provider-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769051
Original-Revision: 27931854ced33febe2c9635b4816ab54243b862b
GitOrigin-RevId: 5efdf3ffd33f9ed2071089fa2ba939bd80925314
Change-Id: If02ea0f55ce8a930470d7952d7bd1f6499184815