[roll] Roll fuchsia [libdriver-integration-test] Remove thread and multiplexing
Make a separate async connection to the controller instead of
spawning a new thread.

This has a good chance of fixing a flake where the device was
torn down, because now the controller call is happening on the
same thread, meaning it is synchronized to the other calls in the test.

Original-Bug: 119548
Multiply: libdriver-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793082
Original-Revision: 059316e9a1d06a3dd93dc2ae882cbaccede93ca6
GitOrigin-RevId: 5a95ce78811aa3d58bfd4ea7916963dfe9c9bbc7
Change-Id: Idd484e7d310a545b64d5f7e57fcc654f5724196c