[roll] Roll llvm-project/libc [libc] fix strtointmax tests
The strtointmax tests weren't running because they depend on the same
shared logic as strtol tests and the target for that was being defined
after the inttypes targets, causing them to be skipped. This patch moves
the inttypes test directory later in the order so that the shared tests
are defined before inttypes tests try to use them.

Reviewed By: sivachandra, lntue

Differential Revision: https://reviews.llvm.org/D142515

GitOrigin-RevId: 209b18249e100b9afa8bbb65f4dc73a654ac1ad5
Original-Revision: 8d83550385455800c2c979e573459b6a6bf6df56
Change-Id: I736d5f15733cf096db9fcf849a73b8ca4dccaab6