[roll] Roll fuchsia [kernel][dlog] Fix alignment of DLog::data_
mvanotti@ pointed out that UBSan noticed some unaligned access of
objects in data_.  This field is an array of bytes that holds records
that start with a dlog_header_t object so we must ensure it's aligned
as 4 or more to avoid undefined behavior.

Update two debuglog test cases to memcpy out of the fifo to avoid the
possibility of unaligned access.

Original-Bug: 105063
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791804
Original-Revision: bb356759818e2c65cd388327a0b7f1fbb5503787
GitOrigin-RevId: 7de6c9d40a782a62ea4c728305f398d9dd70d42b
Change-Id: Idaab31624a0d7dcf4c898de789395a6b2b14edbb