[roll] Roll fuchsia [base-resolver] Use MAC to authenticate contexts
This allows the base-resolver to trust that it created the incoming
contexts and therefore that the contexts refer to base packages
or subpackages of base packages (i.e. it prevents clients from
forging contexts in an attempt to make base-resolver resolve a
subpackage of a non-base package).

Original-Bug: 119496
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792548
Original-Revision: 354fc59407367538f35f2fd7a0bf5a10b5e86ca6
GitOrigin-RevId: f06aa6ad15b5c2bac27877542877b648d15fc0b4
Change-Id: I1a0b2c7f6bbdab8d4fd0a412937367554e413337