[roll] Roll fuchsia [magma] Copy ZirconPlatformConnection to sys_driver_cpp
The MSD C++ API has a new way of sending notifications from the driver
to sys_driver that will require a redesign of the interface to
ZirconPlatformConnection. Since sys_driver_cpp is allowed to depend on
zircon types, it makes sense to copy ZirconPlatformConnection into it
(renaming it to ZirconConnection) to make the process of redesigning the
interface simpler.

In the future this will also help reduce the complexity of some of the
lifetime management, since MagmaSystemConnection will be able to talk
more directly about the async loop and how messages on that are handled.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793132
Original-Revision: 693dd21441d42b1163cf3b57261b41c18c330a07
GitOrigin-RevId: b840ad9fb4b22b98785d1b57be5fc727c1b1fa4d
Change-Id: I9a0afa7bb4f9c5a72b3dbf86df0b7b9cb7062e9a