[roll] Roll fuchsia [gnsdk] Add _hlcpp suffix alias for fidl libraries
Fidl libraries in the GN SDK currently just use the target name
directly for hlcpp bindings, but the new pattern being used in-tree
is to add a suffix specifying which kind of binding it is. This
change adds an alias for the `_hlcpp` suffix in order to allow OOT
users to switch over before the unified bindings are added.

When the unified binding versions are added, the target without a
suffix will cease to exist.

Original-Bug: 115033
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793421
Original-Revision: eb6e721092d766ec7593927e87e58bbf3694ae9b
GitOrigin-RevId: 8c50d7061ce26c456b296e2b4118fc437fe56997
Change-Id: I7deddcb8033f44594e31d1c59ad4d1f26c0f7d59