[roll] Roll fuchsia [cros-es] Migrate core-tests to safer teardown
This CL migrates chromiumos-ec-core-tests to use a single-threaded
dispatcher loop. The motivation for this change is that
|component::OutgoingDirectory| is thread-unsafe and should
not reference a multi-threaded dispatcher. Doing so will
often lead to flakes. If the |component::OutgoingDirectory|
is being destroyed, _while_ there's an in-flight FIDL request
on the internal VFS object, there's a risk for use-after-free
failures.

The pattern used here, with |PerformBlockingWork| is safe and
compatible with multi-threaded code.

Original-Bug: 118385
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791472
Original-Revision: 9b34b3f6d68659ca8d24d4de0040650ae9865f5d
GitOrigin-RevId: f323cde8f490cc270c02637f03912dcdf4311ab5
Change-Id: I79892e138ebd3392d4947e124d89657ab4b95eea