[roll] Roll infra/recipes [api_level_roller] Introduce roller for supported API levels
This roller attempts to roll the list of supported API levels in a
config repo, effectively mirroring the list of supported API levels in
fuchsia.git. This file is taken as-is (a JSON file called
platform_version.json) in fuchsia.git as the source-of-truth, and copied
over into the config repo.

Using Copybara alone isn't feasible since we want to regenerate the
configuration files (which potentially affect the names and
configurations of builders which depend on the supported API levels).

Original-Bug: 119267
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/793400
Original-Revision: c856603342f6453195476daa9b5b8f33f7f61435
GitOrigin-RevId: f2c34c6295c109abec70936d858457cdacc8a2ec
Change-Id: I1ecfcc227717991cb85ccd0e1929ccc796bfb8a6