[roll] Roll fuchsia [alc5663] Migrate acl5663-test  to safer teardown
This CL migrates acl5663-test to use a single-threaded
dispatcher loop. The motivation for this change is that
|component::OutgoingDirectory| is thread-unsafe and should
not reference a multi-threaded dispatcher. Doing so will
often lead to flakes. If the |component::OutgoingDirectory|
is being destroyed, _while_ there's an in-flight FIDL request
on the internal VFS object, there's a risk for use-after-free
failures.

The pattern used here, with |PerformBlockingWork| is safe and
compatible with multi-threaded code.

Original-Bug: 118385
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791473
Original-Revision: dd04f452210b2d752473bc03e493fcea92ff0f6d
GitOrigin-RevId: 6b130896118feb3a419b5bda858ed84c5e9919d3
Change-Id: I12b15fca95462cc6f25fd1d3685b3e61d2bc8b25