[roll] Roll fuchsia [fx] Add prebuilts to $PATH when running commands
To ensure that fx scripts which uses a python3 shebang use the prebuilt
Python included in the checkout instead of the system-installed Python,
which can't be assumed to be a specific version, or even to exist at
all.

This lets us remove a hack from `fx sync-to` that ensured that it always
used the vendored Python prebuilt.

The fx environment cannot be made completely hermetic because many
commands depend on random environment variables, but we can at least
ensure that the $PATH prioritizes prebuilts that are included in the
checkout.

I also merged the `prebuilt.sh` library into `platform.sh` because the
two files overlapped a lot and were used for similar purposes.

Original-Bug: 79905
Original-Bug: 120294
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793360
Original-Revision: 78ba51fa3690ecb2d939711b126a9c42a1e31892
GitOrigin-RevId: ed546d0db281fc1c01101ff3998a4be07721a900
Change-Id: Ia611804264795b745ddbb6e84ffa1aef4fb487dd