[roll] Roll third_party/pigweed pw_build: Ensure TERM is set by pw-wrap-ninja when invoking Ninja
pw-wrap-ninja relies on scraping terminal output from Ninja.
Specifically, it relies on the feature of Ninja where it outputs a
status line both when edge starts running and when an edge finishes
running.

Ninja only has this behavior when it detects that it's running
in a "smart terminal", the logic for which is: stdout must be connected
to a tty *and* the TERM environment variable is set *and* the TERM
environment variable is not 'dumb'.

pw-wrap-ninja always runs Ninja with a pty to satisfy the first part of
the condition, but if it's running without TERM set, Ninja will think
it's in a dumb terminal, and only outputs the "edge finished" line,
which confuses the script (it makes it look like the edge started and
finished running immediately).

This commit rectifies this behavior by ensuring that it sets TERM to
vt100 if no TERM is set. We don't *always* set it, because we would want
to suppress subcommands' colors if we're not actually running in a real
smart terminal (e.g. in a CI job).

Tested: Manually unset TERM. Invoked pw-wrap-ninja, ensured that it was able
  to properly detect edge start/finish, and any output had the terminal
  codes expected.
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126920
Original-Revision: dde2137657998af111f051d619657fbb233fc882
GitOrigin-RevId: 1dda608b482954ba94d57c97c82a4b06070eb2ef
Change-Id: I810fffc056303d1a4c0bb3ff672f3370abfa8a61