[roll] Roll fuchsia [superproject] Roll third_party/pigweed pw_build: Ensure TERM is set by pw-wrap-ninja when invoking Ninja
pw-wrap-ninja relies on scraping terminal output from Ninja.
Specifically, it relies on the feature of Ninja where it outputs a
status line both when edge starts running and when an edge finishes
running.

Ninja only has this behavior when it detects that it's running
in a "smart terminal", the logic for which is: stdout must be connected
to a tty *and* the TERM environment variable is set *and* the TERM
environment variable is not 'dumb'.

pw-wrap-ninja always runs Ninja with a pty to satisfy the first part of
the condition, but if it's running without TERM set, Ninja will think
it's in a dumb terminal, and only outputs the "edge finished" line,
which confuses the script (it makes it look like the edge started and
finished running immediately).

This commit rectifies this behavior by ensuring that it sets TERM to
vt100 if no TERM is set. We don't *always* set it, because we would want
to suppress subcommands' colors if we're not actually running in a real
smart terminal (e.g. in a CI job).

Tested: Manually unset TERM. Invoked pw-wrap-ninja, ensured that it was able
  to properly detect edge start/finish, and any output had the terminal
  codes expected.
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126920
Original-Revision: dde2137657998af111f051d619657fbb233fc882
GitOrigin-RevId: a1d0b44856a579a975bf54c0f5ec2d07653d7614
Roller-URL: https://ci.chromium.org/b/8790794124794482337
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793916
Original-Revision: 753d8595342801fc15624b53eeb2a0cffdd5d671
Change-Id: I535b4f183146cfe0c1ab8b06a1925d750595358f