[roll] Roll fuchsia [ti-tca6408a] Migrate ti-tca6408a to safer teardown
This CL migrates ti-tca6408a-test to use a single-threaded
dispatcher loop. The motivation for this change is that
|component::OutgoingDirectory| is thread-unsafe and should
not reference a multi-threaded dispatcher. Doing so will
often lead to flakes. If the |component::OutgoingDirectory|
is being destroyed, _while_ there's an in-flight FIDL request
on the internal VFS object, there's a risk for use-after-free
failures.

The pattern used here, with |PerformBlockingWork| is safe and
compatible with multi-threaded code.

Original-Bug: 118385
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791474
Original-Revision: 6560a73412fc6c985285c5037fb2f8ac0a8be137
GitOrigin-RevId: 675289130b8394a52a104f0b5c4f62af4832a0b1
Change-Id: Ia0f3444c7750f2e5032a7794d0d0782b3d89d089