[roll] Roll fuchsia [component][fuzzing] Rename Execute to TryOne
Prior to this CL, the word "execute" had become fairly overloaded in the
context of fuzzing. This CL renames the `Execute` method of the
`fuchsia.fuzzer.Controller` protocol to `TryOne` and updates its
implementations and invocations. This better matches the `ffx fuzz`
plugin, and `TryOne(Input)` is more descriptive of what this workflow
actually does.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715746
Original-Revision: 2d12aa941af724616406b53330717584d51d83ca
GitOrigin-RevId: 300dbd66949735b9b9b06c4f47f0eaafa00e0c8f
Change-Id: I30e6e194165c6028bb2e4c05e77a283c7f526f62