[roll] Roll fuchsia [session] Deflake session-manager-integration-test
This test was failing because the graphical-session was trying to
connect to the Scenic protocol but it was not routed.

The test doesn't verify any session behavior, only that it was
launched, so launch the simpler hello-world-session instead.

Original-Fixed: 120541
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794077
Original-Revision: 9693f38ee532bd6da92b3f45da0240af0d7e6959
GitOrigin-RevId: 6040bec4cc2004b62f306e69a012d5eb78cd0edc
Change-Id: I840bc13eb46563ec026da53128ebf2bbf5e03cf6