[roll] Roll third_party/pigweed pw_presubmit: Invoke ninja via pw-wrap-ninja, handle SIGTERM
This commit switches the standard pw_presubmit 'ninja' helper to invoke
Ninja via pw-wrap-ninja, and enable the flag to log action start/end
times.

Additionally, it adds (and uses) an option to the 'call' helper to
propagate a SIGTERM to the subprocess. This is helpful because CI may
terminate pw_presubmit with a SIGTERM if it times out, giving it a
chance to do cleanup. When pw-wrap-ninja receives a SIGTERM, it prints
out any actions it was running. Combined, this makes it easy to see
what's running when a Ninja step in a CI job times out.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126936
Original-Revision: f19e9283a36eebcd0e992f82df0799218948af33
GitOrigin-RevId: 496b126c28217fe71ac6665794d38c08b3b6a02e
Change-Id: I9d44c599ca5126fcb6d441dc9fa0cee2ed5b00d1