[roll] Roll fuchsia [superproject] Roll third_party/pigweed pw_presubmit: Invoke ninja via pw-wrap-ninja, handle SIGTERM
This commit switches the standard pw_presubmit 'ninja' helper to invoke
Ninja via pw-wrap-ninja, and enable the flag to log action start/end
times.

Additionally, it adds (and uses) an option to the 'call' helper to
propagate a SIGTERM to the subprocess. This is helpful because CI may
terminate pw_presubmit with a SIGTERM if it times out, giving it a
chance to do cleanup. When pw-wrap-ninja receives a SIGTERM, it prints
out any actions it was running. Combined, this makes it easy to see
what's running when a Ninja step in a CI job times out.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126936
Original-Revision: f19e9283a36eebcd0e992f82df0799218948af33
GitOrigin-RevId: a251b76e8593aebc74c4cec0c0a5141bad8ab063
Roller-URL: https://ci.chromium.org/b/8790764082931813153
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794260
Original-Revision: fe571542df7bfd39d6c6338a48ea87831e118500
Change-Id: Ife9581a5ac90938914fce7b841a67876542c90c3