[roll] Roll fuchsia [bazel_sdk] merge the starlark rules with the generated SDK
The fx build build/bazel:generate_fuchsia_sdk_repository command is executed by infrastructure to build the bazel sdk and release it to CIPD. The current release does not include the starlark rules and so cannot be used by OOT customers. This change includes the fuchsia and cipd rules so that we can consume this in downstream repos.

This changes is a temporary step to unblocking the release of the fuchsia_sdk rules. The longer term solution will be to invert this relationship and have the generated rules get copied into rules_fuchsia directory. The end result should still have the same shape.

BUG: 114228
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794090
Original-Revision: b173674011c20d5b9f08fbd098fca4f94bb64413
GitOrigin-RevId: 8c050712c84c0117fe63a7b42ca357e5d84f1607
Change-Id: I9d29fce37e70f518b52818c7f4b4e88929f1ffeb