[roll] Roll fuchsia [ffx-test] Remove timeout to check ffx network problems.
We added this timeout to check network problems with ffx but this makes code complicated and error prone. Currently due to this timeout tests
are killed after 1 minute. To make the functionality correct we will
need to pass this timeout everywhere where we make overnet calls which is not feasible and will make the code more complex.

So remove the timeout and only check it once at the start. If we face
network issues with ffx, ffx should handle them internally.

BUG=120384

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793762
Original-Revision: fadcd3cf1042ca56cbdd9e8fd2821c1a530521ba
GitOrigin-RevId: a2bd4005c6eceb34017b9923c5b424df4a4af53a
Change-Id: I871f4913dc8690b669ff86096145ca5b42f9af0e