[roll] Roll fuchsia [starnix] Ignore CLONE_VM when used with CLONE_FORK.
We do not currently have the ability to share address spaces between
processes. This is required to implement clone(CLONE_VM) and the address
space cloning behavior specified by vfork().

By ignoring CLONE_VM when used without CLONE_THREAD (i.e. when making a
child process), this instead performs a copy-on-write clone like fork()
does. This implementation is explicitly permitted by the POSIX
specification as called out in the vfork() man page and the
implementations we have seen that use CLONE_VM are all OK with this
behavior. This flag is mainly used to avoid creating a copy of the
address space when doing a fork because it will be immediately replaced
by an exec() call.

Original-Bug: 114813
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794674
Original-Revision: 40461627029d2a2b736e45770cb1665297c86a1b
GitOrigin-RevId: b94c4cd6d5e33e88315d336ebc904bcac04890a1
Change-Id: Id7d703bc54f06af02a5ba2a72aa960b75cc453f8