[roll] Roll fuchsia [base-resolver] true-base-resolver performs zero-variant fallback
true-base-resolver requires that all base packages, as described
by the static_packages manifest of the system_image package, have
a variant of zero. Clients are allowed to specify or omit the variant.

This mirrors pkg-resolver's behavior and is necessary until the
concept of variant is removed.

Original-Bug: 119496
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794675
Original-Revision: f065ae64f63937e7823cf65c224defa90830374b
GitOrigin-RevId: b2b0c28acb8a816a23fc17da438c13b397bdd535
Change-Id: I22fde8398b27852815fac4fd215aa1cdba8359b1