[roll] Roll fuchsia [doc-checker] Forcing line number correction on Headings
This adds correcting the line number being reported in errors
by matching headings to line numbers in the markdown file.

This is unfortunately needed since markdown is "loose" in its
parsing but the output of the parser is "correct". This means
it correctly parses lists without leading newlines, but the parsed
output gives no indication of it being missing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793141
Original-Revision: cb6f7e1e296df7f4024704fa177835e4902aa895
GitOrigin-RevId: fea416960a68c755d78b080b22de318d2aebde25
Change-Id: I82b10f1df533f32accd5bb9d884bbfe31d5d51a4