[roll] Roll fuchsia [block][abr] Add ABR shim driver
Add a driver that binds to the 'misc' partition on Pinecrest and sets
the one-shot recovery flag when rebooting to recovery or bootloader.
Block ops on this device are passed through to the parent until the
suspend hook is called, at which point the ABR logic takes over if
needed.

Original-Bug: 115458
Original-Bug: b/255495213
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791820
Original-Revision: d8e6c2e98cb3897895f81d6dec354b543edd54fe
GitOrigin-RevId: b7f6637a5ae31d782dad2e8be4a8656f56b875ec
Change-Id: I54f1105bf25f890f51f161f6522ceb7364017321