[roll] Roll fuchsia [a11y][test] Migrates to the new AddLocalChild API
The migration is using the same idea as in fxr/794259.

A fxl::WeakPtr is created in the test fixture, and initialized in
the LocalComponentImpl so that the test fixture can reach into
the LocalComponentImpl's state.

The appropriate init code was changed to wire this through, and
some waits have been set to account for lazy initialization of
the local component.

Original-Bug: 120923
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793578
Original-Revision: 78f59b8d2a7da40b23128e9875fd0980634adf39
GitOrigin-RevId: 23cc1dc3f8271f6b283044da61b54ca19aeb0c14
Change-Id: I13ce7e6ad32e268ee36bd8eeb6883a62588a6b92