[roll] Roll fuchsia [ffx][fho] `Arc` the `Injector` instead of `Box`
Even with the Injector boxed, its lifetime is pretty heavily
constrained, preventing any kind of deferred action on things like
setting up proxies.

This puts it in an Arc so that we can hold on to it for a bit longer and
add a `Deferred` handler (in an upcoming change list) that allows tools
to make decisions about what parts of their initialization struct they
actually need.

Original-Bug: 120550
Original-Bug: 120881
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794440
Original-Revision: 99865919a60ce53b0e2ccd6909cb8befcb2a2f4d
GitOrigin-RevId: 50afa857ef943aebed0eb63127cf921a75b27d9b
Change-Id: Ic65f450e833d522f03007f2a330b78c2441345d7