[roll] Roll fuchsia [netstack] Improve interface watcher logging
A number of the panics and some of the recent logging that was added
are printing addresses as bytes rather than in their human-friendly
forms. Correct all of these, as that might also interfere with
anonymization.

Print all valid-until and preferred-until values as a duration since
boot (which is the monotonic clock's epoch).

Add tests for the Stringer impls.

Original-Bug: 119988
Test: fx test --test-filter '*String*' netstack-gotest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778775
Original-Revision: ba6fd673e8e87e35263865ded3db1c0adc00a9bd
GitOrigin-RevId: e5f71a98f0f034fa90262d1a6b8fd5006aed0800
Change-Id: I871f320650eae8f634b3ed021be27ce34939411d