[roll] Roll fuchsia [pinecrest][emmc] Make abr-shim bind to misc partition and update paver
Add metadata to the eMMC device to make abr-shim bind to the misc
partition in place of block core. abr-shim intercepts calls to reboot to
recovery and uses libabr to set the one-shot recovery flag.

Update the paver so that it uses the mapped GUID to find the misc
partition instead of the GUID on disk.

Original-Bug: 115458
Original-Bug: b/255495213
Test: dm reboot-bootloader and dm reboot-recovery work on Pinecrest
Test: fx test paver-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791821
Original-Revision: c4f8d1784d28679e6461717c1de41746e0f06ea4
GitOrigin-RevId: afde60ed747695c0391c7d8c09ef77cda24a50d3
Change-Id: I2781858e51361acd4013985c84778b33fe89f5f5