[roll] Roll fuchsia [bazel] Speed up sandboxed py_binary() invocations.
This CL changes the python runtime definition to expose
far less files in the sandbox whenever a py_binary() action
needs to run during the build.

This is done by moving 4900+ python module files into a
single zip archive, then launching the python3 interpreter
binary with a PYTHONPATH value that points to the archive
(yes, this works!).

Benefits:

1) The sandbox setup and teardown for actions that run a py_binary()
   will be considerably simpler and faster (3 files instead of ~4940 !!).

2) Bazel aqueries of said actions are _much_ more readable since
   the list of inputs of each action is dramatically shortened.

This also works with remote builders, though it does not reduce
the amount of data that is sent (apart from a much smaller list
of files).

For now, the archive is built by calling the host 'zip' file
within a repository rule. We may consider providing the archive
as part of python3 prebuilt CIPD package instead, but this would
require adjusting our GN build rules as well first, and may have
a noticeable impact on non-sandbox Ninja execution.

An alternative is to add a new CIPD package that provides
//prebuilt/third_party/zip/{host_tag}/zip for all platforms
(e.g. to support Windows in the future).

Original-Fixed: 118192
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795082
Original-Revision: ad8783aa162e927543bbf869056afdffb26c8527
GitOrigin-RevId: 72c95fb84adeb8df0d0afc3154fb97924d479992
Change-Id: Ib0ad93c29a2f5a57a14f9f5d5c15264ae9ed42fb