[roll] Roll fuchsia [iwlwifi][uprev] introduce new device data structure
This is a bulky change for uprev. Mainly this is to refactor
with the new device data structure. But it turns out we need
to change the related code as well.

The changes had been reviewed and LGTMed on the following
CLs:

  fxr/732703, fxr/732704, fxr/732705, fxr/732707, fxr/732708
  fxr/732709, fxr/732710, fxr/732711, fxr/732712, fxr/732713
  fxr/787703, fxr/789173, fxr/732716

Original-Bug: fxb/106535, fxb/106615, fxb/106613, fxb/106535, fxb/103548
Tested: Not breaking AX201 device.
        Passed ready2go script with Atlas and NUC.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795033
Original-Revision: 378fc17a54497d5fcc4ab1b98bfa6f3ee9ff2d97
GitOrigin-RevId: fc793524595ca76bd93a2e1548e491761206019b
Change-Id: Ib3f290bf3173a3b84093734c7e731c5648e5672d