[roll] Roll fuchsia [reachability] Check for DNS activity in eventloop
Query a known domain "www.gstatic.com" with a cadence of 300
seconds to determine whether DNS is active on the system.
Confirm that the response is not none and is a non-empty
vector. Will be used in a future CL to derive DNS activity.

Test success of the lookup_ip call and each of the possible
error types (fidl error, timeout error, lookup error).

Original-Bug: 119352
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793917
Original-Revision: 7e29d94a86578a1ca4753fe86d218d82864ec5c2
GitOrigin-RevId: f3d1e6bed385ce1533438b566c6556bb1a174e7c
Change-Id: I3742c85e6b997064b481ac2dab38944ca8f9493f