[roll] Roll fuchsia [bt][avrcp-tg] Report no song position if playback rate is unavailable
Currently, the playback rate is initialized as a non-null, stopped
playback rate. It is valid for the rate to not exist, in which case the
song position should be None. Store the playback rate as an optional to
account for this.

Test: fx test bt-avrcp-target-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795512
Original-Revision: 420e9fdd76b1ade2b514d6e2fc33916d1c5711e6
GitOrigin-RevId: ab6e6d3bc8b65e013b912855e4ba60c5ebd1209d
Change-Id: I6b0270d3444e734c27e38bed40c792e9d2e17afc