[roll] Roll fuchsia [debugger] Fix number separator parsing.
In Rust code, expressions with "__0" in it would get confusing parsings.
It turns out we were parsing the string as a floating-point number
because '_' is a digit separator in Rust.

Disallows digit separators (also for C which uses ') in the first
position. Add tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795281
Original-Revision: 4894b814ff7caaf5eda5c104eec15ca36c3af925
GitOrigin-RevId: 5d5ce57a8a36eeadc497fea19a8de5dece906238
Change-Id: I19169263576a1317a09b9b4b48ea908692f7af54