[roll] Roll fuchsia [testrunner][ffx] Set timeout on `target wait`.
The default timeout is 2 min, but the default test timeout is 5 min, so
if we retry `ffx target wait` 3 times, we'll hit the test timeout before
we can exit with a fatal error. We should set the timeout to something
smaller so that we can finish the connection attempts before hitting the
timeout.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795723
Original-Revision: 96bf01915bb595f222de195a914902c051d3e8ae
GitOrigin-RevId: 4928af3ca4af80626b008853d510ea1e20f733f7
Change-Id: I27ee89a22f7f13d41acad60b9308da04b7b8af20