[roll] Roll fuchsia [starnix] Introduce abstraction for file I/Os
This CL introduce the notion of InputBuffer and OutputBuffer for file
read and writes.

This simplifies tests, as they do not need to create memory mapping to
execute read and write on files.

This will also be used to write a FIDL file server for starnix
FileObject.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793974
Original-Revision: a9c49f19788c6ba842c5ccc033dd318ea33a19c6
GitOrigin-RevId: 95e01d15b02cc1b1dc6ab81825b405b079233d82
Change-Id: Iccfb8fd566a6abde7da148ff9fbcfee10db48326