[roll] Roll fuchsia [storage][perf] Add --load-blobs-for-tracing flag
When trying to collect a trace immediately after modifying a filesystem
or a benchmark, the start of the trace will be polluted with downloading
the new blobs, writing the blobs to blobfs, and then paging the blobs
back in. Running the benchmark component once with the
--load-blobs-for-tracing flag before running it again with tracing
enabled will remove most of the blob loading from the start of the
trace.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795126
Original-Revision: 35cb99ba2b5bbd0c84fbe2c3a9b71ecfbe3a95f6
GitOrigin-RevId: 54884c07bba78a2346dc90d148d30f2493685e9f
Change-Id: I99fe3506b29e400f2cd6bdda9566c6191a872343