[roll] Roll fuchsia [driver_host] Controller.Rebind returns ZX_ERR_NOT_FOUND
Before this CL the Rebind API would hang forever if it did not cause a
driver to be bound to the device.

Fixing this involves:
- Updating BindDevice to return ZX_ERR_NOT_FOUND if no drivers are
  found.
- Setting the rebind callback before calling rebind.
- Calling the rebind callback in the path where no drivers were found.

The ZX_ERR_NOT_FOUND error is not returned to end clients of Rebind,
as from their perspective the Rebind was successful.

Original-Fixed: 121144
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795602
Original-Revision: 890ff4f88c26a6a0a12e78775b7b2b1e86796321
GitOrigin-RevId: 5577d04b4c643bde0a2c453502650a66802fa124
Change-Id: I50ab21d78136395e203bdf1c7df8d04cde38e028